chg: [stix2 import] Temporary rework stix2 to misp script should now be ready to replace the original stix2 to misp script
