+++ app/files/scripts/stix2/stix2misp.py
@@ -23,50 +23,44 @@ import time
 import io
 import re
 import stix2
-from stix2misp_mapping import *
+import stix2misp_mapping
 from collections import defaultdict
+from copy import deepcopy
+from pathlib import Path
+_misp_dir = Path(os.path.realpath(__file__)).parents[4]
+_misp_objects_path = _misp_dir / 'app' / 'files' / 'misp-objects' / 'objects'
+_pymisp_dir = _misp_dir / 'PyMISP'
+with open(_pymisp_dir / 'pymisp' / 'data' / 'describeTypes.json', 'r') as f:
+    _misp_types = json.loads(f.read())['result'].get('types')
+from pymisp import MISPEvent, MISPObject, MISPAttribute, PyMISPInvalidFormat
 
-_MISP_dir = "/".join([p for p in os.path.dirname(os.path.realpath(__file__)).split('/')[:-4]])
-_PyMISP_dir = '{_MISP_dir}/PyMISP'.format(_MISP_dir=_MISP_dir)
-_MISP_objects_path = '{_MISP_dir}/app/files/misp-objects/objects'.format(_MISP_dir=_MISP_dir)
-sys.path.append(_PyMISP_dir)
-from pymisp.mispevent import MISPEvent, MISPObject, MISPAttribute
-from pymisp.exceptions import PyMISPInvalidFormat
-TAG_REGEX = re.compile(r"\(.+\) .+ = .+")
-special_parsing = ('relationship', 'report', 'galaxy', 'marking-definition')
-galaxy_types = ('attack-pattern', 'intrusion-set', 'malware', 'threat-actor', 'tool')
-with open('{_PyMISP_dir}/pymisp/data/describeTypes.json'.format(_PyMISP_dir=_PyMISP_dir), 'r') as f:
-    misp_types = json.loads(f.read())['result'].get('types')
 
 class StixParser():
-    def __init__(self):
-        super(StixParser, self).__init__()
-        self.misp_event = MISPEvent()
-        self.misp_event['Galaxy'] = []
+    _galaxy_types = ('intrusion-set', 'malware', 'threat-actor', 'tool')
+    _stix2misp_mapping = {'marking-definition': '_load_marking',
+                          'relationship': '_load_relationship',
+                          'report': '_load_report',
+                          'indicator': '_parse_indicator',
+                          'observed-data': '_parse_observable',
+                          'identity': '_load_identity'}
+    _stix2misp_mapping.update({galaxy_type: '_load_galaxy' for galaxy_type in _galaxy_types})
+    _special_mapping = {'attack-pattern': 'parse_attack_pattern',
+                        'course-of-action': 'parse_course_of_action',
+                        'vulnerability': 'parse_vulnerability'}
+    _timeline_mapping = {'indicator': ('valid_from', 'valid_until'),
+                         'observed-data': ('first_observed', 'last_observed')}
 
-    def load_data(self, filename, version, event, args):
-        self.filename = filename
-        self.stix_version = version
-        self.event = defaultdict(dict)
+    def __init__(self):
+        super().__init__()
+        self.misp_event = MISPEvent()
         self.relationship = defaultdict(list)
-        self.event['relationship'] = defaultdict(list)
-        mapping = {'custom_object': '_load_custom',
-                   'marking-definition': '_load_marking',
-                   'relationship': '_load_relationship',
-                   'report': '_load_report'}
-        mapping.update({object_type: '_load_usual_object' for object_type in ('indicator', 'observed-data', 'vulnerability', 'course-of-action')})
-        mapping.update({galaxy_type: '_load_galaxy' for galaxy_type in galaxy_types})
-        for parsed_object in event.objects:
-            try:
-                object_type = parsed_object._type
-            except AttributeError:
-                object_type = parsed_object['type']
-            if object_type.startswith('x-misp-object'):
-                object_type = 'custom_object'
-            try:
-                getattr(self, mapping[object_type])(parsed_object)
-            except KeyError:
-                self._load_usual_object(parsed_object)
+        self.tags = set()
+        self.galaxy = {}
+        self.marking_definition = {}
+
+    def handler(self, event, filename, args):
+        self.filename = filename
+        self.stix_version = f"STIX {event['spec_version'] if event.get('spec_version') else '2.1'}"
         try:
             event_distribution = args[0]
             if not isinstance(event_distribution, int):
@@ -81,546 +75,425 @@ class StixParser():
                 attribute_distribution = int(attribute_distribution) if attribute_distribution.isdigit() else 5
         except IndexError:
             attribute_distribution = 5
-        self.synonyms_to_tag_names = args[2] if len(args) > 2 else '/var/www/MISP/app/files/scripts/synonymsToTagNames.json'
-        self.misp_event.distribution = event_distribution
-        self._attribute_distribution = attribute_distribution
-        self.tags = set()
+        self._synonyms_to_tag_names = args[2] if len(args) > 2 else '/var/www/MISP/app/files/scripts/synonymsToTagNames.json'
+        self.parse_event(event.objects)
 
-    def _load_custom(self, parsed_object):
-        self.event['custom_object'][parsed_object['id'].split('--')[1]] = parsed_object
+    def _load_galaxy(self, galaxy):
+        self.galaxy[galaxy['id'].split('--')[1]] = {'tag_names': self.parse_galaxy(galaxy), 'used': False}
 
-    def _load_galaxy(self, parsed_object):
+    def _load_identity(self, identity):
         try:
-            self.galaxy[parsed_object['id'].split('--')[1]] = {'object': parsed_object, 'used': False}
+            self.identity[identity['id'].split('--')[1]] = identity['name']
         except AttributeError:
-            self.galaxy = {parsed_object['id'].split('--')[1]: {'object': parsed_object, 'used': False}}
+            self.identity = {identity['id'].split('--')[1]: identity['name']}
 
-    def _load_marking(self, parsed_object):
+    def _load_marking(self, marking):
+        tag = self.parse_marking(marking)
+        self.marking_definition[marking['id'].split('--')[1]] = {'object': tag, 'used': False}
+
+    def _load_relationship(self, relationship):
+        target_uuid = relationship.target_ref.split('--')[1]
+        reference = (target_uuid, relationship.relationship_type)
+        source_uuid = relationship.source_ref.split('--')[1]
+        self.relationship[source_uuid].append(reference)
+
+    def _load_report(self, report):
         try:
-            self.marking_definition[parsed_object['id'].split('--')[1]] = {'object': parsed_object, 'used': False}
+            self.report[report['id'].split('--')[1]] = report
         except AttributeError:
-            self.marking_definition = {parsed_object['id'].split('--')[1]: {'object': parsed_object, 'used': False}}
+            self.report = {report['id'].split('--')[1]: report}
 
-    def _load_relationship(self, parsed_object):
-        self.relationship[parsed_object.source_ref.split('--')[1]].append(parsed_object)
+    def _load_synonyms_to_tag_names(self):
+        with open(self._synonyms_to_tag_names, 'rt', encoding='utf-8') as f:
+            synonyms_to_tag_names = json.loads(f.read())
+        self._synonyms_to_tag_names = synonyms_to_tag_names
 
-    def _load_report(self, parsed_object):
-        try:
-            self.report[parsed_object['id'].split('--')[1]] = parsed_object
-        except AttributeError:
-            self.report = {parsed_object['id'].split('--')[1]: parsed_object}
-
-    def _load_usual_object(self, parsed_object):
-        self.event[parsed_object._type][parsed_object['id'].split('--')[1]] = parsed_object
-
-    def general_handler(self):
-        self.outputname = '{}.stix2'.format(self.filename)
-        self.build_from_STIX_with_report() if hasattr(self, 'report') else self.build_from_STIX_without_report()
-        self.set_distribution()
-        if hasattr(self, 'galaxy'):
-            for galaxy in self.galaxy.values():
-                if galaxy['used'] == False:
-                    self.tags.update(self.parse_galaxy(galaxy['object']))
-        if hasattr(self, 'marking_definition'):
-            for marking in self.marking_definition.values():
-                if marking['used'] == False:
-                    try:
-                        self.misp_event.add_tag(self.parse_marking(marking['object']))
-                    except PyMISPInvalidFormat:
-                        continue
-        if self.tags:
-            for tag in self.tags:
-                self.misp_event.add_tag(tag)
-
-    def build_from_STIX_with_report(self):
-        report_attributes = defaultdict(set)
-        for ruuid, report in self.report.items():
-            try:
-                report_attributes['orgs'].add(report.created_by_ref.split('--')[1])
-            except AttributeError:
-                pass
-            report_attributes['name'].add(report.name)
-            if report.get('published'):
-                report_attributes['published'].add(report.published)
-            if 'labels' in report:
-                report_attributes['labels'].update([l for l in report.labels])
-            if 'external_references' in report:
-                self.add_links(report.external_references)
-            for ref in report.object_refs:
-                object_type, uuid = ref.split('--')
-                if object_type not in special_parsing and object_type not in galaxy_types:
-                    try:
-                        object2parse = self.event[object_type][uuid]
-                        self.parsing_process(object2parse, object_type)
-                    except KeyError:
-                        continue
-        if len(report_attributes['orgs']) == 1:
-            identity = self.event['identity'][report_attributes['orgs'].pop()]
-            self.misp_event['Org'] = {'name': identity['name']}
-        if len(report_attributes['published']) == 1:
-            self.misp_event.publish_timestamp = self.getTimestampfromDate(report_attributes['published'].pop())
-        if len(report_attributes['name']) == 1:
-            self.misp_event.info = report_attributes['name'].pop()
-        else:
-            self.misp_event.info = "Imported with MISP import script for {}.".format(self.stix_version)
-        for l in report_attributes['labels']:
-            self.misp_event.add_tag(l)
-
-    def build_from_STIX_without_report(self):
-        for object_type, objects in self.event.items():
-            for _object in objects.values():
-                self.parsing_process(_object, object_type)
-        self.misp_event.info = "Imported with MISP import script for {}.".format(self.stix_version)
-
-    def set_distribution(self):
-        for attribute in self.misp_event.attributes:
-            attribute.distribution = self._attribute_distribution
-        for misp_object in self.misp_event.objects:
-            misp_object.distribution = self._attribute_distribution
-            for attribute in misp_object.attributes:
-                attribute.distribution = self._attribute_distribution
-
-    def saveFile(self):
-        eventDict = self.misp_event.to_json()
-        outputfile = '{}.stix2'.format(self.filename)
-        with open(outputfile, 'w') as f:
-            f.write(eventDict)
-
-    def add_links(self, refs):
-        for e in refs:
-            link = {"type": "link"}
-            comment = e.get('source_name')
-            try:
-                comment = comment.split('url - ')[1]
-            except IndexError:
-                pass
-            if comment:
-                link['comment'] = comment
-            link['value'] = e.get('url')
-            self.misp_event.add_attribute(**link)
-
-    @staticmethod
-    def getTimestampfromDate(stix_date):
-        try:
-            return int(stix_date.timestamp())
-        except AttributeError:
-            return int(time.mktime(time.strptime(stix_date.split('+')[0], "%Y-%m-%d %H:%M:%S")))
-
-    @staticmethod
-    def get_misp_type(labels):
-        return labels[0].split('=')[1][1:-1]
-
-    @staticmethod
-    def get_misp_category(labels):
-        return labels[1].split('=')[1][1:-1]
+    def save_file(self):
+        event = self.misp_event.to_json()
+        with open(f'{self.filename}.stix2', 'wt', encoding='utf-8') as f:
+            f.write(event)
 
     ################################################################################
     ##                 PARSING FUNCTIONS USED BY BOTH SUBCLASSES.                 ##
     ################################################################################
 
-    def add_tag_in_attribute(self, attribute, marking_refs):
-        attribute = self.pyMISPify(attribute)
-        for marking in marking_refs:
-            marking_uuid = marking.split('--')[1]
-            marking = self.marking_definition[marking_uuid]
-            tag = self.parse_marking(marking['object'])
-            if tag is not None:
-                attribute.add_tag(self.parse_marking(marking['object']))
-            marking['used'] = True
-        return attribute
-
-    @staticmethod
-    def append_email_attribute(_type, value, to_ids):
-        mapping = email_mapping[_type]
-        return {'type': mapping['type'], 'object_relation': mapping['relation'], 'value': value, 'to_ids': to_ids}
-
-    def attributes_from_asn_observable(self, objects):
-        attributes = []
-        for _object in objects.values():
-            attributes.extend(self.fill_observable_attributes(_object, asn_mapping))
-        return attributes
-
-    def attributes_from_domain_ip_observable(self, objects):
-        attributes = []
-        for _object in objects.values():
-            mapping = domain_ip_mapping[_object.type]
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': _object.value, 'to_ids': False})
-        return attributes
-
-    def attributes_from_file_observable(self, _object, data=None):
-        attributes = []
-        md5 = None
-        if hasattr(_object, 'hashes'):
-            if 'MD5' in _object.hashes:
-                md5 = _object.hashes['MD5']
-            for h_type, h_value in _object.hashes.items():
-                h_type = h_type.lower().replace('-', '')
-                attributes.append({'type': h_type, 'object_relation': h_type,
-                                   'value': h_value, 'to_ids': False})
-        attributes.extend(self.fill_observable_attributes(_object, file_mapping))
-        if data is not None and md5 and hasattr(_object, 'name'):
-            attributes.append({'type': 'malware-sample', 'object_relation': 'malware-sample',
-                               'value': '{}|{}'.format(_object.name, md5), 'data': data})
-        return attributes
-
-    @staticmethod
-    def attributes_from_file_pattern(types, values):
-        attributes = []
-        for type_, value in zip(types, values):
-            if 'hashes' in type_:
-                hash_type = type_.split('.')[1].strip("'").replace('-', '').lower()
-                attributes.append({'type': hash_type, 'value': value,
-                                   'object_relation': hash_type, 'to_ids': True})
-            else:
-                try:
-                    mapping = file_mapping[type_]
-                    attributes.append({'type': mapping['type'], 'value': value,
-                                       'object_relation': mapping['relation'], 'to_ids': True})
-                except KeyError:
-                    continue
-        return attributes
-
-    def attributes_from_network_traffic(self, objects, name=None):
-        network_traffic, references = self.fetch_network_traffic_objects_and_references(objects)
-        attributes = self.fill_observable_attributes(network_traffic, network_traffic_mapping)
-        if name is not None:
-            mapping = network_socket_types
-            for protocol in network_traffic.protocols:
-                try:
-                    layer = connection_protocols[protocol]
-                    attributes.append({'type': 'text', 'value': protocol, 'to_ids': False,
-                                       'object_relation': 'layer{}-protocol'.format(layer)})
-                except KeyError:
-                    continue
-        elif hasattr(network_traffic, 'extensions') and network_traffic.extensions:
-            extension_type, extension_value = list(network_traffic.extensions.items())[0]
-            name = network_traffic_extensions[extension_type]
-            attributes.extend(self.parse_socket_extension(extension_value))
-            mapping = network_traffic_references_mapping['with_extensions']
-        else:
-            name = 'ip-port'
-            mapping = network_traffic_references_mapping['without_extensions']
-        attributes.extend(self.parse_network_traffic_references(references, network_traffic, mapping))
-        if references:
-            attributes.extend(self.parse_remaining_references(references, mapping))
-        return attributes, name
-
-    def attributes_from_process_observable(self, objects):
-        main_process = None
-        for _object in objects.values():
-            if hasattr(_object, 'parent_ref') or hasattr(_object, 'child_refs'):
-                main_process = _object
-                break
-        if main_process:
-            attributes = self.fill_observable_attributes(main_process, process_mapping)
-            for refs in ('parent_ref', 'child_refs'):
-                if hasattr(main_process, refs):
-                    attributes.extend([self.parse_reference_process(objects[ref], refs) for ref in getattr(main_process, refs)])
-            return attributes
-        return [result for _object in objects.values() for result in self.fill_observable_attributes(_object, process_mapping)]
-
-    @staticmethod
-    def parse_reference_process(process, ref_type):
-        field = 'pid'
-        mapping = process_mapping[field]
-        relation = '{}-{}'.format(ref_type.split('_')[0], mapping['relation'])
-        return {'type': mapping['type'], 'object_relation': relation, 'value': getattr(process, field), 'to_ids': False}
-
-    def attributes_from_regkey_observable(self, _object):
-        attributes = []
-        for key, value in _object.items():
-            if key in regkey_mapping:
-                mapping = regkey_mapping[key]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': value.replace('\\\\', '\\'), 'to_ids': False})
-        if 'values' in _object:
-            attributes.extend(self.fill_observable_attributes(_object.values[0], regkey_mapping))
-        return attributes
-
-    @staticmethod
-    def attributes_from_url_observable(objects):
-        attributes = []
-        for value in objects.values():
-            if isinstance(value, (stix2.URL, stix2.DomainName)):
-                mapping = url_mapping[value._type]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': value.value, 'to_ids': False})
-            elif isinstance(value, stix2.NetworkTraffic):
-                mapping = url_mapping[value._type]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': value.dst_port, 'to_ids': False})
-        return attributes
-
-    def attributes_from_user_account_observable(self, observable):
-        observable = observable['0']
-        attributes = self.fill_user_account_observable_attributes(observable)
-        if 'extensions' in observable and 'unix-account-ext' in observable['extensions']:
-            extension = observable['extensions']['unix-account-ext']
-            if 'groups' in extension:
-                for group in extension['groups']:
-                    attributes.append({'type': 'text', 'object_relation': 'group',
-                                       'to_ids': False, 'disable_correlation': True,
-                                       'value': group})
-        attributes.extend(self.fill_user_account_observable_attributes(extension))
-        return attributes
-
-    def attributes_from_user_account_pattern(self, pattern):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            p_value = p_value[1:-1]
-            if "extensions.'unix-account-ext'" in p_type:
-                relation = p_type.split('.')[-1]
-                if 'groups' in relation:
-                    attributes.append({'type': 'text', 'object_relation': 'group',
-                                       'disable_correlation': True, 'value': p_value})
-                    continue
-            else:
-                relation = p_type.split(':')[1]
-            if relation in user_account_mapping:
-                attribute = {'value': p_value}
-                attribute.update(user_account_mapping[relation])
-                attributes.append(attribute)
-        return attributes
-
-    def attributes_from_x509_observable(self, objects):
-        _object = objects['0']
-        attributes = self.fill_observable_attributes(_object.hashes, x509_mapping) if (hasattr(_object, 'hashes') and _object.hashes) else []
-        attributes.extend(self.fill_observable_attributes(_object, x509_mapping))
-        return attributes
-
-    @staticmethod
-    def split_file_observable(objects):
-        data = None
-        for value in objects.values():
-            if isinstance(value, stix2.Artifact):
-                data = value.payload_bin
-            elif isinstance(value, stix2.Directory):
-                data = value
-            elif isinstance(value, stix2.File):
-                file = value
-        return file, data
-
-    @staticmethod
-    def fetch_network_traffic_objects_and_references(objects):
-        references = {}
-        for key, value in objects.items():
-            if isinstance(value, (stix2.DomainName, stix2.IPv4Address, stix2.IPv6Address)):
-                references[key] = value
-            elif isinstance(value, stix2.NetworkTraffic):
-                network_traffic = value
-        return network_traffic, references
-
-    @staticmethod
-    def fill_user_account_observable_attributes(observable):
-        attributes = []
-        for key, value in observable.items():
-            if key in user_account_mapping:
-                attribute = {'to_ids': False, 'value': value}
-                attribute.update(user_account_mapping[key])
-                attributes.append(attribute)
-        return attributes
-
-    def handle_object_relationship(self, misp_object, uuid):
-        for reference in self.relationship[uuid]:
-            target = reference.target_ref.split('--')[1]
-            if target not in self.galaxy:
-                misp_object.add_reference(target, reference.relationship_type)
-
-    def handle_single_attribute(self, attribute, uuid=None):
-        if uuid is not None:
-            if uuid in self.relationship:
-                galaxies = []
-                for reference in self.relationship[uuid]:
-                    target = reference.target_ref.split('--')[1]
-                    if target in self.galaxy:
-                        galaxy = self.galaxy[target]
-                        galaxies.append(self.parse_galaxy(galaxy['object'])[0])
-                        galaxy['used'] = True
-                if galaxies:
-                    attribute['Tag'] = galaxies
-        self.misp_event.add_attribute(**attribute)
-
-    def parse_complex_fields_observable_email(self, objects, to_ids):
-        attributes = []
-        addresses, files, message = self.split_observable_email_parts(objects)
-        if 'from_ref' in message:
-            from_ref = 'from_ref'
-            attributes.append(self.append_email_attribute(from_ref, addresses[message.pop(from_ref)], to_ids))
-        for ref in ('to_refs', 'cc_refs'):
-            if ref in message:
-                attributes.extend([self.append_email_attribute(ref, addresses[item], to_ids) for item in message.pop(ref)])
-        if 'body_multipart' in message:
-            brr = 'body_raw_ref'
-            attributes.extend([self.append_email_attribute('body_multipart', files[f[brr]], to_ids) for f in message.pop('body_multipart') if brr in f])
-        if 'additional_header_fields' in message:
-            for field_key, field_value in message.pop('additional_header_fields').items():
-                if field_key == 'Reply-To':
-                    attributes.extend([self.append_email_attribute(field_key, reply_to, to_ids) for reply_to in field_value])
-                else:
-                    attributes.append(self.append_email_attribute(field_key, field_value, to_ids))
-        return attributes, message
-
-    def parse_course_of_action(self, o):
-        misp_object = MISPObject('course-of-action', misp_objects_path_custom=_MISP_objects_path)
-        if 'name' in o:
-            attribute = {'type': 'text', 'object_relation': 'name', 'value': o.get('name')}
-            misp_object.add_attribute(**attribute)
-        else:
-            return
-        if 'description' in o:
-            attribute = {'type': 'text', 'object_relation': 'description', 'value': o.get('description')}
-            misp_object.add_attribute(**attribute)
-        self.misp_event.add_object(**misp_object)
-
-    @staticmethod
-    def __parse_network_traffic_reference(ref_object, ref, mapping):
-        origin = ref.split('_')[0]
-        misp_type, relation = mapping[ref_object._type]
-        return {'type': misp_type.format(origin), 'object_relation': relation.format(origin),
-                'to_ids': False, 'value': ref_object.value}
-
-    def parse_network_traffic_references(self, objects, network_traffic, mapping):
-        attributes= []
-        for ref in ('src_ref', 'dst_ref'):
-            if hasattr(network_traffic, ref):
-                ref_object = objects.pop(getattr(network_traffic, ref))
-                attributes.append(self.__parse_network_traffic_reference(ref_object, ref, mapping))
-        for refs in ('src_refs', 'dst_refs'):
-            if hasattr(network_traffic, refs):
-                for ref in getattr(network_traffic, refs):
-                    ref_object = objects.pop(ref)
-                    attributes.append(self.__parse_network_traffic_reference(ref_object, refs, mapping))
-        return attributes
-
-    def parse_pe(self, extension):
-        pe = MISPObject('pe', misp_objects_path_custom=_MISP_objects_path)
-        self.fill_object_attributes_observable(pe, pe_mapping, extension)
-        for section in extension['sections']:
-            pe_section = MISPObject('pe-section', misp_objects_path_custom=_MISP_objects_path)
-            if 'hashes' in section:
-                for h_type, h_value in section['hashes'].items():
-                    h_type = h_type.lower().replace('-', '')
-                    pe_section.add_attribute(**{'type': h_type, 'object_relation': h_type,
-                                                'value': h_value, 'to_ids': False})
-            self.fill_object_attributes_observable(pe_section, pe_section_mapping, section)
-            pe.add_reference(pe_section.uuid, 'includes')
-            self.misp_event.add_object(**pe_section)
-        self.misp_event.add_object(**pe)
-        return pe.uuid
-
-    @staticmethod
-    def parse_remaining_references(references, mapping):
-        attributes = []
-        for reference in references.values():
-            misp_type, relation = mapping[reference._type]
-            attributes.append({'type': misp_type, 'object_relation': relation,
-                               'to_ids': False, 'value': reference.value})
-        return attributes
-
-    @staticmethod
-    def parse_socket_extension(extension):
-        attributes = []
-        for element in extension:
-            try:
-                mapping = network_traffic_mapping[element]
-            except KeyError:
-                continue
-            if element in ('is_listening', 'is_blocking'):
-                attribute_value = element.split('_')[1]
-            else:
-                attribute_value = extension[element]
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': attribute_value})
-        return attributes
-
-    @staticmethod
-    def pyMISPify(attribute_dict):
+    def create_attribute_with_tag(self, attribute_dict, marking_refs):
         attribute = MISPAttribute()
         attribute.from_dict(**attribute_dict)
+        try:
+            self.marking_refs[attribute.uuid] = (marking.split('--')[1] for marking in marking_refs)
+        except AttributeError:
+            self.marking_refs = {attribute.uuid: (marking.split('--')[1] for marking in marking_refs)}
         return attribute
 
     @staticmethod
-    def split_observable_email_parts(observable):
-        addresses = {}
-        files = {}
-        message = None
-        for o_key, o_dict in observable.items():
-            if isinstance(o_dict, stix2.EmailAddress):
-                addresses[o_key] = o_dict.value
-            elif isinstance(o_dict, stix2.EmailMessage):
-                message = dict(o_dict)
-            elif isinstance(o_dict, stix2.File):
-                files[o_key] = o_dict.name
-        return addresses, files, message
+    def _parse_email_body(body, references):
+        attributes = []
+        for body_multipart in body:
+            reference = references.pop(body_multipart['body_raw_ref'])
+            feature = body_multipart['content_disposition'].split(';')[0]
+            if feature in stix2misp_mapping.email_references_mapping:
+                attribute = deepcopy(stix2misp_mapping.email_references_mapping[feature])
+            else:
+                print(f'Unknown content disposition in the following email body: {body_multipart}', file=sys.stderr)
+                continue
+            if isinstance(reference, stix2.Artifact):
+                attribute.update({
+                    'value': body_multipart['content_disposition'].split('=')[-1].strip("'"),
+                    'data': reference.payload_bin,
+                    'to_ids': False
+                })
+            else:
+                attribute.update({
+                    'value': reference.name,
+                    'to_ids': False
+                })
+            attributes.append(attribute)
+        return attributes
+
+    @staticmethod
+    def _parse_email_references(email_message, references):
+        attributes = []
+        if hasattr(email_message, 'from_ref'):
+            reference = references.pop(email_message.from_ref)
+            attribute = {
+                'value': reference.value,
+                'to_ids': False
+            }
+            attribute.update(stix2misp_mapping.email_references_mapping['from_ref'])
+            attributes.append(attribute)
+        for feature in ('to_refs', 'cc_refs'):
+            if hasattr(email_message, feature):
+                for ref_id in getattr(email_message, feature):
+                    reference = references.pop(ref_id)
+                    attribute = {
+                        'value': reference.value,
+                        'to_ids': False
+                    }
+                    attribute.update(stix2misp_mapping.email_references_mapping[feature])
+                    attributes.append(attribute)
+        return attributes
+
+    def parse_galaxies(self):
+        for galaxy in self.galaxy.values():
+            if not galaxy['used']:
+                for tag_name in galaxy['tag_names']:
+                    self.tags.add(tag_name)
+
+    @staticmethod
+    def _parse_network_connection_reference(feature_type, feature, value):
+        if feature == 'type':
+            return {type: value.format(feature_type) for type, value in stix2misp_mapping.network_traffic_references_mapping[value].items()}
+        return {feature: value}
+
+    @staticmethod
+    def _parse_network_traffic_protocol(protocol):
+        return {'type': 'text', 'value': protocol, 'to_ids': False,
+                'object_relation': f'layer{stix2misp_mapping.connection_protocols[protocol]}-protocol'}
+
+    @staticmethod
+    def _parse_observable_reference(reference, mapping, feature=None):
+        attribute = {
+            'value': reference.value,
+            'to_ids': False
+        }
+        if feature is not None:
+            attribute.update({key: value.format(feature) for key, value in getattr(stix2misp_mapping, mapping)[reference._type].items()})
+            return attribute
+        attribute.update({key: value for key, value in getattr(stix2misp_mapping, mapping)[reference._type].items()})
+        return attribute
+
+    def parse_pe(self, extension):
+        pe_object = MISPObject('pe', misp_objects_path_custom=_misp_objects_path)
+        self.fill_misp_object(pe_object, extension, 'pe_mapping')
+        for section in extension['sections']:
+            section_object = MISPObject('pe-section', misp_objects_path_custom=_misp_objects_path)
+            self.fill_misp_object(section_object, section, 'pe_section_mapping')
+            if hasattr(section, 'hashes'):
+                self.fill_misp_object(section_object, section.hashes, 'pe_section_mapping')
+            pe_object.add_reference(section_object.uuid, 'includes')
+            self.misp_event.add_object(**section_object)
+        self.misp_event.add_object(**pe_object)
+        return pe_object.uuid
+
+    def parse_relationships(self):
+        attribute_uuids = tuple(attribute.uuid for attribute in self.misp_event.attributes)
+        object_uuids = tuple(object.uuid for object in self.misp_event.objects)
+        for source, references in self.relationship.items():
+            if source in self.galaxy:
+                continue
+            if source in object_uuids:
+                source_object = self.misp_event.get_object_by_uuid(source)
+                for reference in references:
+                    target, reference = reference
+                    if target in attribute_uuids or target in object_uuids:
+                        source_object.add_reference(target, reference)
+            elif source in attribute_uuids:
+                for attribute in self.misp_event.attributes:
+                    if attribute.uuid == source:
+                        for reference in references:
+                            target, reference = reference
+                            if target in self.galaxy:
+                                for tag_name in self.galaxy[target]['tag_names']:
+                                    attribute.add_tag(tag_name)
+                                self.galaxy[target]['used'] = True
+                        break
+
+    def parse_report(self):
+        event_infos = set()
+        if len(self.report) == 1:
+            self.misp_event.uuid = tuple(self.report.keys())[0]
+        for report in self.report.values():
+            if hasattr(report, 'name') and report.name:
+                event_infos.add(report.name)
+            if hasattr(report, 'labels') and report.labels:
+                for label in report.labels:
+                    self.tags.add(label)
+            if hasattr(report, 'object_marking_refs') and report.object_marking_refs:
+                for marking_ref in report.object_marking_refs:
+                    marking_ref = marking_ref.split('--')[1]
+                    try:
+                        self.tags.add(self.marking_definition[marking_ref]['object'])
+                        self.marking_definition[marking_ref]['used'] = True
+                    except KeyError:
+                        continue
+            if hasattr(report, 'external_references'):
+                for reference in report.external_references:
+                    self.misp_event.add_attribute(**{'type': 'link', 'value': reference['url']})
+        if len(event_infos) == 1:
+            self.misp_event.info = event_infos.pop()
+        else:
+            self.misp_event.info = f'Imported with MISP import script for {self.stix_version}'
+
+    @staticmethod
+    def _parse_user_account_groups(groups):
+        attributes = [{'type': 'text', 'object_relation': 'group', 'to_ids': False,
+                       'disable_correlation': True, 'value': group} for group in groups]
+        return attributes
+
+    ################################################################################
+    ##                             UTILITY FUNCTIONS.                             ##
+    ################################################################################
+
+    @staticmethod
+    def _choose_with_priority(container, first_choice, second_choice):
+        return first_choice if first_choice in container else second_choice
+
+    def filter_main_object(self, observable, main_type, test_function='_standard_test_filter'):
+        references = {}
+        main_objects = []
+        for key, value in observable.items():
+            if getattr(self, test_function)(value, main_type):
+                main_objects.append(value)
+            else:
+                references[key] = value
+        if len(main_objects) > 1:
+            print(f'More than one {main_type} objects in this observable: {observable}', file=sys.stderr)
+        return main_objects[0] if main_objects else None, references
+
+    def _get_tag_names_from_synonym(self, name):
+        try:
+            return self._synonyms_to_tag_names[name]
+        except TypeError:
+            self._load_synonyms_to_tag_names()
+            return self._synonyms_to_tag_names[name]
+
+    @staticmethod
+    def getTimestampfromDate(date):
+        try:
+            return int(date.timestamp())
+        except AttributeError:
+            return int(time.mktime(time.strptime(date.split('+')[0], "%Y-%m-%dT%H:%M:%S.%fZ")))
+
+    @staticmethod
+    def _handle_data(data):
+        return io.BytesIO(data.encode())
+
+    @staticmethod
+    def parse_marking(marking):
+        marking_type = marking.definition_type
+        tag = getattr(marking.definition, marking_type)
+        return "{}:{}".format(marking_type, tag)
+
+    def parse_timeline(self, stix_object):
+        misp_object = {'timestamp': self.getTimestampfromDate(stix_object.modified)}
+        try:
+            first, last = self._timeline_mapping[stix_object._type]
+            first_seen = getattr(stix_object, first)
+            if stix_object.created != first_seen and stix_object.modified != first_seen:
+                misp_object['first_seen'] = first_seen
+                if hasattr(stix_object, last):
+                    misp_object['last_seen'] = getattr(stix_object, last)
+            elif hasattr(stix_object, last):
+                misp_object.update({'first_seen': first_seen, 'last_seen': getattr(stix_object, last)})
+        except KeyError:
+            pass
+        return misp_object
+
+    @staticmethod
+    def _process_test_filter(value, main_type):
+        _is_main_process = any(feature in value for feature in ('parent_ref', 'child_refs'))
+        return isinstance(value, getattr(stix2, main_type)) and _is_main_process
+
+    @staticmethod
+    def _standard_test_filter(value, main_type):
+        return isinstance(value, getattr(stix2, main_type))
 
 
 class StixFromMISPParser(StixParser):
+    _objects_mapping = {'asn': {'observable': 'parse_asn_observable',
+                                'pattern': 'parse_asn_pattern'},
+                        'credential': {'observable': 'parse_credential_observable',
+                                       'pattern': 'parse_credential_pattern'},
+                        'domain-ip': {'observable': 'parse_domain_ip_observable',
+                                      'pattern': 'parse_domain_ip_pattern'},
+                        'email': {'observable': 'parse_email_observable',
+                                  'pattern': 'parse_email_pattern'},
+                        'file': {'observable': 'parse_file_observable',
+                                 'pattern': 'parse_file_pattern'},
+                        'ip-port': {'observable': 'parse_ip_port_observable',
+                                    'pattern': 'parse_ip_port_pattern'},
+                        'network-connection': {'observable': 'parse_network_connection_observable',
+                                               'pattern': 'parse_network_connection_pattern'},
+                        'network-socket': {'observable': 'parse_network_socket_observable',
+                                           'pattern': 'parse_network_socket_pattern'},
+                        'process': {'observable': 'parse_process_observable',
+                                    'pattern': 'parse_process_pattern'},
+                        'registry-key': {'observable': 'parse_regkey_observable',
+                                         'pattern': 'parse_regkey_pattern'},
+                        'url': {'observable': 'parse_url_observable',
+                                'pattern': 'parse_url_pattern'},
+                        'user-account': {'observable': 'parse_user_account_observable',
+                                         'pattern': 'parse_user_account_pattern'},
+                        'WindowsPEBinaryFile': {'observable': 'parse_pe_observable',
+                                                'pattern': 'parse_pe_pattern'},
+                        'x509': {'observable': 'parse_x509_observable',
+                                 'pattern': 'parse_x509_pattern'}}
+
     def __init__(self):
-        super(StixFromMISPParser, self).__init__()
-        self.objects_mapping = {'asn': {'observable': 'attributes_from_asn_observable', 'pattern': 'pattern_asn'},
-                                'credential': {'observable': 'observable_credential', 'pattern': 'pattern_credential'},
-                                'domain-ip': {'observable': 'attributes_from_domain_ip_observable', 'pattern': 'pattern_domain_ip'},
-                                'email': {'observable': 'observable_email', 'pattern': 'pattern_email'},
-                                'file': {'observable': 'observable_file', 'pattern': 'pattern_file'},
-                                'ip-port': {'observable': 'observable_ip_port', 'pattern': 'pattern_ip_port'},
-                                'network-connection': {'observable': 'observable_connection', 'pattern': 'pattern_connection'},
-                                'network-socket': {'observable': 'observable_socket', 'pattern': 'pattern_socket'},
-                                'process': {'observable': 'attributes_from_process_observable', 'pattern': 'pattern_process'},
-                                'registry-key': {'observable': 'attributes_from_regkey_observable', 'pattern': 'pattern_regkey'},
-                                'url': {'observable': 'attributes_from_url_observable', 'pattern': 'pattern_url'},
-                                'user-account': {'observable': 'attributes_from_user_account_observable',
-                                                 'pattern': 'attributes_from_user_account_pattern'},
-                                'WindowsPEBinaryFile': {'observable': 'observable_pe', 'pattern': 'pattern_pe'},
-                                'x509': {'observable': 'attributes_from_x509_observable', 'pattern': 'pattern_x509'}}
-        self.object_from_refs = {'course-of-action': 'parse_MISP_course_of_action', 'vulnerability': 'parse_vulnerability',
-                                 'custom_object': 'parse_custom'}
-        self.object_from_refs.update(dict.fromkeys(['indicator', 'observed-data'], 'parse_usual_object'))
-        self.attributes_fetcher_mapping = {'indicator': 'fetch_attributes_from_indicator',
-                                           'observed-data': 'fetch_attributes_from_observable',
-                                           'vulnerability': 'fetch_attributes_from_vulnerability'}
+        super().__init__()
+        self._stix2misp_mapping.update({'custom_object': '_parse_custom'})
+        self._stix2misp_mapping.update({special_type: '_parse_undefined' for special_type in ('attack-pattern', 'course-of-action', 'vulnerability')})
 
-    def handler(self):
-        self.general_handler()
+    def parse_event(self, stix_objects):
+        for stix_object in stix_objects:
+            object_type = stix_object['type']
+            if object_type.startswith('x-misp-object'):
+                object_type = 'custom_object'
+            if object_type in self._stix2misp_mapping:
+                getattr(self, self._stix2misp_mapping[object_type])(stix_object)
+            else:
+                print(f'not found: {object_type}', file=sys.stderr)
+        if self.relationship:
+            self.parse_relationships()
+        if self.galaxy:
+            self.parse_galaxies()
+        if hasattr(self, 'report'):
+            self.parse_report()
+        if self.marking_definition:
+            for marking_definition in self.marking_definition.values():
+                if not marking_definition['used']:
+                    self.tags.add(marking_definition['object'])
+        if self.tags:
+            for tag in self.tags:
+                self.misp_event.add_tag(tag)
 
-    def parsing_process(self, object2parse, object_type):
-        labels = object2parse.get('labels')
-        getattr(self, self.object_from_refs[object_type])(object2parse, labels)
+    def _parse_custom(self, custom):
+        if 'from_object' in custom['labels']:
+            self.parse_custom_object(custom)
+        else:
+            self.parse_custom_attribute(custom)
+
+    def _parse_indicator(self, indicator):
+        if 'from_object' in indicator['labels']:
+            self.parse_indicator_object(indicator)
+        else:
+            self.parse_indicator_attribute(indicator)
+
+    def _parse_observable(self, observable):
+        if 'from_object' in observable['labels']:
+            self.parse_observable_object(observable)
+        else:
+            self.parse_observable_attribute(observable)
+
+    def _parse_undefined(self, stix_object):
+        if any(label.startswith('misp-galaxy:') for label in stix_object.get('labels', [])):
+            self._load_galaxy(stix_object)
+        else:
+            getattr(self, self._special_mapping[stix_object._type])(stix_object)
 
     ################################################################################
     ##                             PARSING FUNCTIONS.                             ##
     ################################################################################
 
-    def parse_usual_object(self, o, labels):
-        if 'from_object' in labels:
-            self.parse_object(o, labels)
-        else:
-            self.parse_attribute(o, labels)
+    def fill_misp_object(self, misp_object, stix_object, mapping,
+                         to_call='_fill_observable_object_attribute'):
+        for feature, value in stix_object.items():
+            if feature not in getattr(stix2misp_mapping, mapping):
+                if feature.startswith('x_misp_'):
+                    attribute = self.parse_custom_property(feature)
+                    if isinstance(value, list):
+                        self._fill_misp_object_from_list(misp_object, attribute, value)
+                        continue
+                else:
+                    continue
+            else:
+                attribute = deepcopy(getattr(stix2misp_mapping, mapping)[feature])
+            attribute.update(getattr(self, to_call)(feature, value))
+            misp_object.add_attribute(**attribute)
 
-    def parse_galaxy(self, galaxy):
-        return [galaxy['labels'][1]]
+    @staticmethod
+    def _fill_misp_object_from_list(misp_object, mapping, values):
+        for value in values:
+            attribute = {'value': value}
+            attribute.update(mapping)
+            misp_object.add_attribute(**attribute)
 
-    def parse_MISP_course_of_action(self, o, _):
-        self.parse_course_of_action(o)
+    def parse_attack_pattern(self, attack_pattern):
+        misp_object, _ = self.create_misp_object(attack_pattern)
+        if hasattr(attack_pattern, 'external_references'):
+            for reference in attack_pattern.external_references:
+                value = reference['external_id'].split('-')[1] if reference['source_name'] == 'capec' else reference['url']
+                misp_object.add_attribute(**{
+                    'type': 'text', 'object_relation': 'id',
+                    'value': value
+                })
+        self.fill_misp_object(misp_object, attack_pattern, 'attack_pattern_mapping',
+                              '_fill_observable_object_attribute')
+        self.misp_event.add_object(**misp_object)
 
-    def parse_custom(self, o, labels):
-        if 'from_object' in labels:
-            self.parse_custom_object(o, labels)
-        else:
-            self.parse_custom_attribute(o, labels)
+    def parse_course_of_action(self, course_of_action):
+        misp_object, _ = self.create_misp_object(course_of_action)
+        self.fill_misp_object(misp_object, course_of_action, 'course_of_action_mapping',
+                              '_fill_observable_object_attribute')
+        self.misp_event.add_object(**misp_object)
 
-    def parse_custom_object(self, o, labels):
-        name = o['type'].split('x-misp-object-')[1]
-        misp_object = MISPObject(name, misp_objects_path_custom=_MISP_objects_path)
-        misp_object.timestamp = self.getTimestampfromDate(o['x_misp_timestamp'])
-        misp_object.uuid = o['id'].split('--')[1]
+    def parse_custom_attribute(self, custom):
+        attribute_type = custom['type'].split('x-misp-object-')[1]
+        if attribute_type not in _misp_types:
+            replacement = ' ' if attribute_type == 'named-pipe' else '|'
+            attribute_type = attribute_type.replace('-', replacement)
+        attribute = {'type': attribute_type,
+                     'timestamp': self.getTimestampfromDate(custom['modified']),
+                     'to_ids': bool(custom['labels'][1].split('=')[1]),
+                     'value': custom['x_misp_value'],
+                     'category': self.get_misp_category(custom['labels']),
+                     'uuid': custom['id'].split('--')[1]}
+        if custom.get('object_marking_refs'):
+            attribute = self.create_attribute_with_tag(attribute, custom['object_marking_refs'])
+        self.misp_event.add_attribute(**attribute)
+
+    def parse_custom_object(self, custom):
+        name = custom['type'].split('x-misp-object-')[1]
+        misp_object = MISPObject(name, misp_objects_path_custom=_misp_objects_path)
+        misp_object.timestamp = self.getTimestampfromDate(custom['modified'])
+        misp_object.uuid = custom['id'].split('--')[1]
         try:
-            misp_object.category = o['category']
+            misp_object.category = custom['category']
         except KeyError:
-            misp_object.category = self.get_misp_category(labels)
+            misp_object.category = self.get_misp_category(custom['labels'])
         attributes = []
-        for key, value in o['x_misp_values'].items():
+        for key, value in custom['x_misp_values'].items():
             attribute_type, object_relation = key.split('_')
             if isinstance(value, list):
                 for single_value in value:
@@ -631,927 +504,1619 @@ class StixFromMISPParser(StixParser):
                                              'object_relation': object_relation})
         self.misp_event.add_object(**misp_object)
 
-    def parse_custom_attribute(self, o, labels):
-        attribute_type = o['type'].split('x-misp-object-')[1]
-        if attribute_type not in misp_types:
-            replacement = ' ' if attribute_type == 'named-pipe' else '|'
-            attribute_type = attribute_type.replace('-', replacement)
-        attribute = {'type': attribute_type,
-                     'timestamp': self.getTimestampfromDate(o['x_misp_timestamp']),
-                     'to_ids': bool(labels[1].split('=')[1]),
-                     'value': o['x_misp_value'],
-                     'category': self.get_misp_category(labels),
-                     'uuid': o['id'].split('--')[1]}
-        if o.get('object_marking_refs'):
-            attribute = self.add_tag_in_attribute(attribute, o['object_marking_refs'])
+    def parse_galaxy(self, galaxy):
+        if hasattr(galaxy, 'labels'):
+            return tuple(label for label in galaxy.labels if label.startswith('misp-galaxy:'))
+        try:
+            return tuple(self._get_tag_names_from_synonym(galaxy.name))
+        except KeyError:
+            print(f'Unknown {galaxy._type} name: {galaxy.name}', file=sys.stderr)
+            return tuple()
+
+    def parse_indicator_attribute(self, indicator):
+        attribute = self.create_attribute_dict(indicator)
+        attribute['to_ids'] = True
+        pattern = indicator.pattern.replace('\\\\', '\\')
+        if attribute['type'] in ('malware-sample', 'attachment'):
+            value, data = self.parse_attribute_pattern_with_data(pattern)
+            attribute.update({feature: value for feature, value in zip(('value', 'data'), (value, io.BytesIO(data.encode())))})
+        else:
+            attribute['value'] = self.parse_attribute_pattern(pattern)
+        if hasattr(indicator, 'object_marking_refs'):
+            attribute = self.create_attribute_with_tag(attribute, indicator.object_marking_refs)
         self.misp_event.add_attribute(**attribute)
 
-    def parse_object(self, o, labels):
-        object_type = self.get_misp_type(labels)
-        name = 'file' if object_type == 'WindowsPEBinaryFile' else object_type
-        object_category = self.get_misp_category(labels)
-        stix_type = o._type
-        misp_object = MISPObject(name, misp_objects_path_custom=_MISP_objects_path)
-        uuid = o.id.split('--')[1]
-        misp_object.uuid = uuid
-        misp_object['meta-category'] = object_category
+    def parse_indicator_object(self, indicator):
+        misp_object, object_type = self.create_misp_object(indicator)
+        pattern = indicator.pattern.replace('\\\\', '\\').strip('[]').split(' AND ')
         try:
-            attributes = getattr(self, self.attributes_fetcher_mapping[stix_type])(o, object_type)
+            attributes = getattr(self, self._objects_mapping[object_type]['pattern'])(pattern)
         except KeyError:
-            print("Unable to map {} object:\n{}".format(stix_type, o), file=sys.stderr)
+            print(f"Unable to map {object_type} object:\n{indicator}", file=sys.stderr)
             return
         if isinstance(attributes, tuple):
-            attributes, pe_uuid = attributes
-            misp_object.add_reference(pe_uuid, 'includes')
+            attributes, target_uuid = attributes
+            misp_object.add_reference(target_uuid, 'includes')
         for attribute in attributes:
             misp_object.add_attribute(**attribute)
-        misp_object.to_ids = (labels[2].split('=')[1][1:-1].lower() == 'true')
-        if uuid in self.relationship:
-            self.handle_object_relationship(misp_object, uuid)
         self.misp_event.add_object(**misp_object)
 
-    def fetch_attributes_from_indicator(self, indicator, object_type):
-        pattern = indicator.pattern.replace('\\\\', '\\').split(' AND ')
-        pattern[0] = pattern[0][1:]
-        pattern[-1] = pattern[-1][:-1]
-        return getattr(self, self.objects_mapping[object_type]['pattern'])(pattern)
-
-    def fetch_attributes_from_observable(self, observable, object_type):
-        observable = observable.objects
-        return getattr(self, self.objects_mapping[object_type]['observable'])(observable)
-
-    def fetch_attributes_from_vulnerability(self, vulnerability, _):
-        attributes = []
-        if vulnerability.get('name'):
-            attributes.append({'type': 'text', 'object_relation': 'id', 'value': vulnerability['name']})
-        if vulnerability.get('description'):
-            attributes.append({'type': 'text', 'object_relation': 'summary', 'value': vulnerability['description']})
-        if vulnerability.get('external_references'):
-            for reference in vulnerability['external_references']:
-                if reference['source_name'] == 'url':
-                    attributes.append({'type': 'link', 'object_relation': 'references', 'value': reference['url']})
-        return attributes
-
-    def parse_attribute(self, o, labels):
-        attribute_uuid = o.id.split('--')[1]
-        attribute_type = self.get_misp_type(labels)
-        attribute_category = self.get_misp_category(labels)
-        attribute = {'uuid': attribute_uuid, 'type': attribute_type, 'category': attribute_category}
-        tags = [{'name': label} for label in labels[3:]]
-        if tags:
-            attribute['Tag'] = tags
-        stix_type = o._type
-        if stix_type == 'vulnerability':
-            value = o.get('name')
-        else:
-            if stix_type == 'indicator':
-                if hasattr(o, 'valid_until'):
-                    org_uuid = o.created_by_ref.split('--')[1]
-                    attribute['Sighting'] = [{'type': '2', 'date_sighting': str(self.getTimestampfromDate(o.valid_until)),
-                                             'Organisation': {'uuid': org_uuid, 'name': self.event['identity'][org_uuid]['name']}}]
-                pattern = o.pattern.replace('\\\\', '\\')
-                value = self.parse_pattern_with_data(pattern) if attribute_type in ('malware-sample', 'attachment') else self.parse_pattern(pattern)
-                attribute['to_ids'] = True
-            else:
-                attribute['timestamp'] = self.getTimestampfromDate(o.get('last_observed'))
-                observable = o.objects
-                try:
-                    value = misp_types_mapping[attribute_type](observable, attribute_type)
-                except Exception:
-                    print('Error with attribute type {}:\n{}'.format(attribute_type, observable), file=sys.stderr)
-                attribute['to_ids'] = False
-        if hasattr(o, 'description'):
-            attribute['comment'] = o.description
+    def parse_observable_attribute(self, observable):
+        attribute = self.create_attribute_dict(observable)
+        attribute['to_ids'] = False
+        objects = observable.objects
+        value = self.parse_single_attribute_observable(objects, attribute['type'])
         if isinstance(value, tuple):
             value, data = value
-            attribute['data'] = io.BytesIO(data.encode())
+            attribute['data'] = data
         attribute['value'] = value
-        if hasattr(o, 'object_marking_refs'):
-            attribute = self.add_tag_in_attribute(attribute, o.object_marking_refs)
-        self.handle_single_attribute(attribute, uuid=attribute_uuid)
+        if hasattr(observable, 'object_marking_refs'):
+            attribute = self.create_attribute_with_tag(attribute, indicator.object_marking_refs)
+        self.misp_event.add_attribute(**attribute)
 
-    def parse_vulnerability(self, vulnerability, labels):
-        if len(labels) > 2:
-            self.parse_usual_object(vulnerability, labels)
+    def parse_observable_object(self, observable):
+        misp_object, object_type = self.create_misp_object(observable)
+        observable_object = observable.objects
+        try:
+            attributes = getattr(self, self._objects_mapping[object_type]['observable'])(observable_object)
+        except KeyError:
+            print(f"Unable to map {object_type} object:\n{observable}", file=sys.stderr)
+            return
+        if isinstance(attributes, tuple):
+            attributes, target_uuid = attributes
+            misp_object.add_reference(target_uuid, 'includes')
+        for attribute in attributes:
+            misp_object.add_attribute(**attribute)
+        self.misp_event.add_object(**misp_object)
+
+    def parse_vulnerability(self, vulnerability):
+        attributes = self.fill_observable_attributes(vulnerability, 'vulnerability_mapping')
+        if hasattr(vulnerability, 'external_references'):
+            for reference in vulnerability.external_references:
+                if reference['source_name'] == 'url':
+                    attributes.append({'type': 'link', 'object_relation': 'references', 'value': reference['url']})
+        if len(attributes) > 1:
+            vulnerability_object, _ = self.create_misp_object(vulnerability)
+            for attribute in attributes:
+                vulnerability_object.add_attribute(**attribute)
+            self.misp_event.add_object(**vulnerability_object)
         else:
-            self.tags.update(self.parse_galaxy(vulnerability))
+            attribute = self.create_attribute_dict(vulnerability)
+            attribute['value'] = attributes[0]['value']
+            self.misp_event.add_attribute(**attribute)
 
-    def observable_connection(self, observable):
-        attributes, _ = self.attributes_from_network_traffic(observable, 'network-connection')
-        return attributes
-
-    def pattern_connection(self, pattern):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            p_value = p_value[1:-1]
-            try:
-                mapping = network_traffic_mapping[p_type]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': p_value})
-            except KeyError:
-                if not p_type.startswith('network-traffic:protocols['):
-                    continue
-                attributes.append({'type': 'text', 'value': p_value,
-                                   'object_relation': 'layer{}-protocol'.format(connection_protocols[p_value])})
-        return attributes
-
-    def observable_credential(self, observable):
-        return self.fill_observable_attributes(observable['0'], credential_mapping)
-
-    def pattern_credential(self, pattern):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            p_type = p_type.split(':')[1]
-            p_value = p_value[1:-1]
-            try:
-                mapping = credential_mapping[p_type]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': p_value})
-            except KeyError:
-                if not p_type.startswith('x_misp_'):
-                    continue
-                attribute_type, relation = p_type.strip('x_misp_').split('_')
-                attributes.append({'type': attribute_type, 'object_relation': relation,
-                                   'value': p_value})
-        return attributes
-
-    def observable_email(self, observable):
-        to_ids = False
-        attributes, message = self.parse_complex_fields_observable_email(observable, to_ids)
-        for m_key, m_value in message.items():
-            try:
-                attributes.append(self.append_email_attribute(m_key, m_value, False))
-            except KeyError:
-                if m_key.startswith("x_misp_attachment_"):
-                    attribute_type, relation = m_key.split("x_misp_")[1].split("_")
-                    attributes.append({'type': attribute_type, 'object_relation': relation, 'to_ids': False,
-                                       'value': m_value['value'], 'data': io.BytesIO(m_value['data'].encode())})
-                elif "x_misp_" in m_key:
-                    attribute_type, relation = m_key.split("x_misp_")[1].split("_")
-                    attributes.append({'type': attribute_type, 'object_relation': relation,
-                                       'value': m_value, 'to_ids': False})
-        return attributes
-
-    def pattern_email(self, pattern):
-        attributes = []
-        attachments = defaultdict(dict)
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            try:
-                attributes.append(self.append_email_attribute(p_type, p_value[1:-1], True))
-            except KeyError:
-                if p_type.startswith("email-message:'x_misp_attachment_"):
-                    relation, field = p_type.split('.')
-                    relation = relation.split(':')[1][1:-1]
-                    attachments[relation][field] = p_value[1:-1]
-                elif "x_misp_" in p_type:
-                    attribute_type, relation = p_type.split("x_misp_")[1][:-1].split("_")
-                    attributes.append({'type': attribute_type, 'object_relation': relation,
-                                       'value': p_value[1:-1], 'to_ids': True})
-        for a_key, a_dict in attachments.items():
-            _, _, attribute_type, relation = a_key.split('_')
-            attributes.append({'type': attribute_type, 'object_relation': relation, 'to_ids': True,
-                               'value': a_dict['value'], 'data': io.BytesIO(a_dict['data'].encode())})
-        return attributes
-
-    def observable_file(self, observable):
-        if len(observable) > 1:
-            file, data = self.split_file_observable(observable)
-            if data is not None:
-                return self.attributes_from_file_observable(file, data)
-        return self.attributes_from_file_observable(observable['0'])
+    ################################################################################
+    ##                        OBSERVABLE PARSING FUNCTIONS                        ##
+    ################################################################################
 
     @staticmethod
-    def pattern_file(pattern):
+    def _define_hash_type(hash_type):
+        if 'sha' in hash_type:
+            return f'SHA-{hash_type.split("sha")[1]}'
+        return hash_type.upper() if hash_type == 'md5' else hash_type
+
+    @staticmethod
+    def _fetch_file_observable(observable_objects):
+        for key, observable in observable_objects.items():
+            if observable['type'] == 'file':
+                return key
+        return '0'
+
+    @staticmethod
+    def _fill_observable_attribute(attribute_type, object_relation, value):
+        return {'type': attribute_type,
+                'object_relation': object_relation,
+                'value': value,
+                'to_ids': False}
+
+    def fill_observable_attributes(self, observable, object_mapping):
         attributes = []
-        malware_sample = {}
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            if p_type == 'artifact:payload_bin':
-                malware_sample['data'] = p_value
-            elif p_type in ("file:name", "file:hashes.'md5'"):
-                try:
-                    mapping = file_mapping[p_type]
-                    attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                       'value': p_value[1:-1], 'to_ids': True})
-                    malware_sample['filename'] = p_value[1:-1]
-                except KeyError:
-                    attributes.append({'type': 'md5', 'object_relation': 'md5',
-                                       'value': p_value[1:-1], 'to_ids': True})
-                    malware_sample['md5'] = p_value[1:-1]
-            elif 'file:hashes.' in p_type:
-                _, h = p_type.split('.')
-                h = h[1:-1]
-                attributes.append({'type': h, 'object_relation': h, 'value': p_value[1:-1]})
+        for key, value in observable.items():
+            if key in getattr(stix2misp_mapping, object_mapping):
+                attribute = deepcopy(getattr(stix2misp_mapping, object_mapping)[key])
+            elif key.startswith('x_misp_'):
+                attribute = self.parse_custom_property(key)
+                if isinstance(value, list):
+                    for single_value in value:
+                        single_attribute = {'value': single_value, 'to_ids': False}
+                        single_attribute.update(attribute)
+                        attributes.append(single_attribute)
+                    continue
             else:
-                try:
-                    mapping = file_mapping[p_type]
-                    attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                       'value': p_value[1:-1], 'to_ids': True})
-                except KeyError:
-                    if "x_misp_" in  p_type:
-                        attribute_type, relation = p_type.split("x_misp_")[1][:-1].split("_")
-                        attributes.append({'type': attribute_type, 'object_relation': relation,
-                                           'value': p_value[1:-1], 'to_ids': True})
-        if 'data' in malware_sample:
-            value = "{}|{}".format(malware_sample['filename'], malware_sample['md5'])
-            attributes.append({'type': 'malware-sample', 'object_relation': 'malware-sample',
-                               'value': value, 'to_ids': True, 'data': io.BytesIO(malware_sample['data'].encode())})
+                continue
+            attribute.update({'value': value, 'to_ids': False})
+            attributes.append(attribute)
         return attributes
 
-    def observable_pe(self, observable):
-        extension = observable['1']['extensions']['windows-pebinary-ext']
-        pe_uuid = self.parse_pe(extension)
-        return self.observable_file(observable), pe_uuid
+    def _handle_multiple_file_fields(self, file):
+        attributes = []
+        for feature, attribute_type in zip(('filename', 'path', 'fullpath'), ('filename', 'text', 'text')):
+            key = f'x_misp_multiple_{feature}'
+            if key in file:
+                attributes.append(self._fill_observable_attribute(attribute_type, feature, file.pop(key)))
+            elif f'{key}s' in file:
+                attributes.extend(self._fill_observable_attribute(attribute_type, feature, value) for value in file.pop(key))
+        attributes.extend(self.fill_observable_attributes(file, 'file_mapping'))
+        return attributes
 
-    def pattern_pe(self, pattern):
+    def parse_asn_observable(self, observable):
+        attributes = []
+        mapping = 'asn_mapping'
+        for observable_object in observable.values():
+            if isinstance(observable_object, stix2.AutonomousSystem):
+                attributes.extend(self.fill_observable_attributes(observable_object, mapping))
+            else:
+                attributes.append(self._parse_observable_reference(observable_object, mapping))
+        return attributes
+
+    def _parse_attachment(self, observable):
+        if len(observable) > 1:
+            return self._parse_name(observable, index='1'), self._parse_payload(observable)
+        return self._parse_name(observable)
+
+    def parse_credential_observable(self, observable):
+        return self.fill_observable_attributes(observable['0'], 'credential_mapping')
+
+    def _parse_domain_ip_attribute(self, observable):
+        return f'{self._parse_value(observable)}|{self._parse_value(observable, index="1")}'
+
+    @staticmethod
+    def parse_domain_ip_observable(observable):
+        attributes = []
+        for object in observable.values():
+            attribute = deepcopy(stix2misp_mapping.domain_ip_mapping[object._type])
+            attribute.update({'value': object.value, 'to_ids': False})
+            attributes.append(attribute)
+        return attributes
+
+    @staticmethod
+    def _parse_email_message(observable, attribute_type):
+        return observable['0'].get(attribute_type.split('-')[1])
+
+    def parse_email_observable(self, observable):
+        email, references = self.filter_main_object(observable, 'EmailMessage')
+        attributes = self.fill_observable_attributes(email, 'email_mapping')
+        if hasattr(email, 'additional_header_fields'):
+            attributes.extend(self.fill_observable_attributes(email.additional_header_fields, 'email_mapping'))
+        attributes.extend(self._parse_email_references(email, references))
+        if hasattr(email, 'body_multipart') and email.body_multipart:
+            attributes.extend(self._parse_email_body(email.body_multipart, references))
+        return attributes
+
+    @staticmethod
+    def _parse_email_reply_to(observable):
+        return observable['0'].additional_header_fields.get('Reply-To')
+
+    def parse_file_observable(self, observable):
+        file, references = self.filter_main_object(observable, 'File')
+        references = {key: {'object': value, 'used': False} for key, value in references.items()}
+        file = {key: value for key, value in file.items()}
+        multiple_fields = any(f'x_misp_multiple_{feature}' in file for feature in ('filename', 'path', 'fullpath'))
+        attributes = self._handle_multiple_file_fields(file) if multiple_fields else self.fill_observable_attributes(file, 'file_mapping')
+        if 'hashes' in file:
+            attributes.extend(self.fill_observable_attributes(file['hashes'], 'file_mapping'))
+        if 'content_ref' in file:
+            reference = references[file['content_ref']]
+            value = f'{reference["object"].name}|{reference["object"].hashes["MD5"]}'
+            attributes.append({'type': 'malware-sample', 'object_relation': 'malware-sample', 'value': value,
+                               'to_ids': False, 'data': reference['object'].payload_bin})
+            reference['used'] = True
+        if 'parent_directory_ref' in file:
+            reference = references[file['parent_directory_ref']]
+            attributes.append({'type': 'text', 'object_relation': 'path',
+                               'value': reference['object'].path, 'to_ids': False})
+            reference['used'] = True
+        for reference in references.values():
+            if not reference['used']:
+                attributes.append({
+                    'type': 'attachment',
+                    'object_relation': 'attachment',
+                    'value': reference['object'].name,
+                    'data': reference['object'].payload_bin,
+                    'to_ids': False
+                })
+        return attributes
+
+    def _parse_filename_hash(self, observable, attribute_type, index='0'):
+        hash_type = attribute_type.split('|')[1]
+        filename = self._parse_name(observable, index=index)
+        hash_value = self._parse_hash(observable, hash_type, index=index)
+        return f'{filename}|{hash_value}'
+
+    def _parse_hash(self, observable, attribute_type, index='0'):
+        hash_type = self._define_hash_type(attribute_type)
+        return observable[index]['hashes'].get(hash_type)
+
+    def parse_ip_port_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable, 'NetworkTraffic')
+        attributes = []
+        for feature in ('src', 'dst'):
+            port = f'{feature}_port'
+            if hasattr(network_traffic, port):
+                attribute = deepcopy(stix2misp_mapping.ip_port_mapping[port])
+                attribute.update({'value': getattr(network_traffic, port), 'to_ids': False})
+                attributes.append(attribute)
+            ref = f'{feature}_ref'
+            if hasattr(network_traffic, ref):
+                attributes.append(self._parse_observable_reference(references.pop(getattr(network_traffic, ref)), 'ip_port_references_mapping', feature))
+        for reference in references.values():
+            attribute = deepcopy(stix2misp_mapping.ip_port_references_mapping[reference._type])
+            attribute.update({'value': reference.value, 'to_ids': False})
+            attributes.append(attribute)
+        return attributes
+
+    def _parse_malware_sample(self, observable):
+        if len(observable) > 1:
+            value = self._parse_filename_hash(observable, 'filename|md5', '1')
+            return value, self._parse_payload(observable)
+        return self._parse_filename_hash(observable, 'filename|md5')
+
+    @staticmethod
+    def _parse_name(observable, index='0'):
+        return observable[index].get('name')
+
+    def _parse_network_attribute(self, observable):
+        port = self._parse_port(observable, index='1')
+        return f'{self._parse_value(observable)}|{port}'
+
+    def parse_network_connection_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable, 'NetworkTraffic')
+        attributes = self._parse_network_traffic(network_traffic, references)
+        if hasattr(network_traffic, 'protocols'):
+            attributes.extend(self._parse_network_traffic_protocol(protocol) for protocol in network_traffic.protocols if protocol in stix2misp_mapping.connection_protocols)
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, 'domain_ip_mapping'))
+        return attributes
+
+    def parse_network_socket_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable, 'NetworkTraffic')
+        attributes = self._parse_network_traffic(network_traffic, references)
+        if hasattr(network_traffic, 'protocols'):
+            attributes.append({'type': 'text', 'object_relation': 'protocol', 'to_ids': False,
+                               'value': network_traffic.protocols[0].strip("'")})
+        if hasattr(network_traffic, 'extensions') and network_traffic.extensions:
+            attributes.extend(self._parse_socket_extension(network_traffic.extensions['socket-ext']))
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, 'domain_ip_mapping'))
+        return attributes
+
+    def _parse_network_traffic(self, network_traffic, references):
+        attributes = []
+        mapping = 'network_traffic_references_mapping'
+        for feature in ('src', 'dst'):
+            port = f'{feature}_port'
+            if hasattr(network_traffic, port):
+                attribute = deepcopy(stix2misp_mapping.network_traffic_mapping[port])
+                attribute.update({'value': getattr(network_traffic, port), 'to_ids': False})
+                attributes.append(attribute)
+            ref = f'{feature}_ref'
+            if hasattr(network_traffic, ref):
+                attributes.append(self._parse_observable_reference(references.pop(getattr(network_traffic, ref)), mapping, feature))
+            if hasattr(network_traffic, f'{ref}s'):
+                for ref in getattr(network_traffic, f'{ref}s'):
+                    attributes.append(self._parse_observable_reference(references.pop(ref), mapping, feature))
+        return attributes
+
+    @staticmethod
+    def _parse_number(observable):
+        return observable['0'].get('number')
+
+    @staticmethod
+    def _parse_payload(observable):
+        return observable['0'].payload_bin
+
+    def parse_pe_observable(self, observable):
+        key = self._fetch_file_observable(observable)
+        extension = observable[key]['extensions']['windows-pebinary-ext']
+        pe_uuid = self.parse_pe(extension)
+        return self.parse_file_observable(observable), pe_uuid
+
+    @staticmethod
+    def _parse_port(observable, index='0'):
+        port_observable = observable[index]
+        return port_observable['src_port'] if 'src_port' in port_observable else port_observable['dst_port']
+
+    def parse_process_observable(self, observable):
+        process, references = self.filter_main_object(observable, 'Process', test_function='_process_test_filter')
+        attributes = self.fill_observable_attributes(process, 'process_mapping')
+        if hasattr(process, 'parent_ref'):
+            attributes.extend(self.fill_observable_attributes(references[process.parent_ref], 'parent_process_reference_mapping'))
+        if hasattr(process, 'child_refs'):
+            for reference in process.child_refs:
+                attributes.extend(self.fill_observable_attributes(references[reference], 'child_process_reference_mapping'))
+        if hasattr(process, 'binary_ref'):
+            reference = references[process.binary_ref]
+            attribute = deepcopy(stix2misp_mapping.process_image_mapping)
+            attribute.update({'value': reference.name, 'to_ids': False})
+            attributes.append(attribute)
+        return attributes
+
+    @staticmethod
+    def _parse_regkey_attribute(observable):
+        return observable['0'].get('key')
+
+    def parse_regkey_observable(self, observable):
+        attributes = []
+        for key, value in observable['0'].items():
+            if key in stix2misp_mapping.regkey_mapping:
+                attribute = deepcopy(stix2misp_mapping.regkey_mapping[key])
+                attribute.update({'value': value.replace('\\\\', '\\'), 'to_ids': False})
+                attributes.append(attribute)
+        if 'values' in observable['0']:
+            attributes.extend(self.fill_observable_attributes(observable['0']['values'][0], 'regkey_mapping'))
+        return attributes
+
+    def _parse_regkey_value(self, observable):
+        regkey = self._parse_regkey_attribute(observable)
+        return f'{regkey}|{observable["0"]["values"][0].get("data")}'
+
+    def parse_single_attribute_observable(self, observable, attribute_type):
+        if attribute_type in stix2misp_mapping.attributes_type_mapping:
+            return getattr(self, stix2misp_mapping.attributes_type_mapping[attribute_type])(observable, attribute_type)
+        return getattr(self, stix2misp_mapping.attributes_mapping[attribute_type])(observable)
+
+    def _parse_socket_extension(self, extension):
+        attributes = []
+        extension = {key: value for key, value in extension.items()}
+        if 'x_misp_text_address_family' in extension:
+            extension.pop('address_family')
+        for element, value in extension.items():
+            if element in stix2misp_mapping.network_socket_extension_mapping:
+                attribute = deepcopy(stix2misp_mapping.network_socket_extension_mapping[element])
+                if element in ('is_listening', 'is_blocking'):
+                    if value is False:
+                        continue
+                    value = element.split('_')[1]
+            elif element.startswith('x_misp_'):
+                attribute = self.parse_custom_property(element)
+            else:
+                continue
+            attribute.update({'value': value, 'to_ids': False})
+            attributes.append(attribute)
+        return attributes
+
+    def parse_url_observable(self, observable):
+        attributes = []
+        for object in observable.values():
+            feature = 'dst_port' if isinstance(object, stix2.NetworkTraffic) else 'value'
+            attribute = deepcopy(stix2misp_mapping.url_mapping[object._type])
+            attribute.update({'value': getattr(object, feature), 'to_ids': False})
+            attributes.append(attribute)
+        return attributes
+
+    def parse_user_account_observable(self, observable):
+        observable = observable['0']
+        attributes = self.fill_observable_attributes(observable, 'user_account_mapping')
+        if 'extensions' in observable and 'unix-account-ext' in observable['extensions']:
+            extension = observable['extensions']['unix-account-ext']
+            if 'groups' in extension:
+                attributes.extend(self._parse_user_account_groups(extension['groups']))
+            attributes.extend(self.fill_observable_attributes(extension, 'user_account_mapping'))
+        return attributes
+
+    @staticmethod
+    def _parse_value(observable, index='0'):
+        return observable[index].get('value')
+
+    def _parse_x509_attribute(self, observable, attribute_type):
+        hash_type = attribute_type.split('-')[-1]
+        return self._parse_hash(observable, hash_type)
+
+    def parse_x509_observable(self, observable):
+        attributes = self.fill_observable_attributes(observable['0'], 'x509_mapping')
+        if hasattr(observable['0'], 'hashes') and observable['0'].hashes:
+            attributes.extend(self.fill_observable_attributes(observable['0'].hashes, 'x509_mapping'))
+        return attributes
+
+    ################################################################################
+    ##                         PATTERN PARSING FUNCTIONS.                         ##
+    ################################################################################
+
+    def fill_pattern_attributes(self, pattern, object_mapping):
+        attributes = []
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            if pattern_type not in getattr(stix2misp_mapping, object_mapping):
+                if 'x_misp_' in pattern_type:
+                    attribute = self.parse_custom_property(pattern_type)
+                    attribute['value'] = pattern_value.strip("'")
+                    attributes.append(attribute)
+                continue
+            attribute = deepcopy(getattr(stix2misp_mapping, object_mapping)[pattern_type])
+            attribute['value'] = pattern_value.strip("'")
+            attributes.append(attribute)
+        return attributes
+
+    def parse_asn_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'asn_mapping')
+
+    def parse_credential_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'credential_mapping')
+
+    def parse_domain_ip_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'domain_ip_mapping')
+
+    def parse_email_pattern(self, pattern):
+        attributes = []
+        attachments = defaultdict(dict)
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            if 'body_multipart' in pattern_type:
+                pattern_type = pattern_type.split('.')
+                feature = 'data' if pattern_type[-1] == 'payload_bin' else 'value'
+                attachments[pattern_type[0][-2]][feature] = pattern_value.strip("'")
+                continue
+            if pattern_type not in stix2misp_mapping.email_mapping:
+                if 'x_misp_' in pattern_type:
+                    attribute = self.parse_custom_property(pattern_type)
+                    attribute['value'] = pattern_value.strip("'")
+                    attributes.append(attribute)
+                continue
+            attribute = deepcopy(stix2misp_mapping.email_mapping[pattern_type])
+            attribute['value'] = pattern_value.strip("'")
+            attributes.append(attribute)
+        for attachment in attachments.values():
+            if 'data' in attachment:
+                attribute = {'type': 'attachment', 'object_relation': 'screenshot', 'data': attachment['data']}
+            else:
+                attribute = {'type': 'email-attachment', 'object_relation': 'attachment'}
+            attribute['value'] = attachment['value']
+            attributes.append(attribute)
+        return attributes
+
+    def parse_file_pattern(self, pattern):
+        attributes = []
+        attachment = {}
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            if pattern_type in stix2misp_mapping.attachment_types:
+                attachment[pattern_type] = pattern_value.strip("'")
+            if pattern_type not in stix2misp_mapping.file_mapping:
+                continue
+            attribute = deepcopy(stix2misp_mapping.file_mapping[pattern_type])
+            attribute['value'] = pattern_value.strip("'")
+            attributes.append(attribute)
+        if 'file:content_ref.payload_bin' in attachment:
+            filename = self._choose_with_priority(attachment, 'file:content_ref.name', 'file:name')
+            md5 = self._choose_with_priority(attachment, "file:content_ref.hashes.'MD5'", "file:hashes.'MD5'")
+            attributes.append({
+                'type': 'malware-sample',
+                'object_relation': 'malware-sample',
+                'value': f'{attachment[filename]}|{attachment[md5]}',
+                'data': attachment['file:content_ref.payload_bin']
+            })
+        if 'artifact:payload_bin' in attachment:
+            attributes.append({
+                'type': 'attachment',
+                'object_relation': 'attachment',
+                'value': attachment['artifact:x_misp_text_name'] if 'artifact:x_misp_text_name' in attachment else attachment['file:name'],
+                'data': attachment['artifact:payload_bin']
+            })
+        return attributes
+
+    def parse_ip_port_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'ip_port_mapping')
+
+    def parse_network_connection_pattern(self, pattern):
+        attributes = []
+        references = defaultdict(dict)
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            if pattern_type not in stix2misp_mapping.network_traffic_mapping:
+                pattern_value = pattern_value.strip("'")
+                if pattern_type.startswith('network-traffic:protocols['):
+                    attributes.append({
+                        'type': 'text', 'value': pattern_value,
+                        'object_relation': f'layer{stix2misp_mapping.connection_protocols[pattern_value]}-protocol'
+                    })
+                elif any(pattern_type.startswith(f'network-traffic:{feature}_ref') for feature in ('src', 'dst')):
+                    feature_type, ref = pattern_type.split(':')[1].split('_')
+                    ref, feature = ref.split('.')
+                    ref = f"{feature_type}_{'0' if ref == 'ref' else ref.strip('ref[]')}"
+                    references[ref].update(self._parse_network_connection_reference(feature_type, feature, pattern_value))
+                continue
+            attribute = deepcopy(stix2misp_mapping.network_traffic_mapping[pattern_type])
+            attribute['value'] = pattern_value.strip("'")
+            attributes.append(attribute)
+        attributes.extend(attribute for attribute in references.values())
+        return attributes
+
+    def parse_network_socket_pattern(self, pattern):
+        attributes = []
+        references = defaultdict(dict)
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            pattern_value = pattern_value.strip("'")
+            if pattern_type not in stix2misp_mapping.network_traffic_mapping:
+                if pattern_type in stix2misp_mapping.network_socket_extension_mapping:
+                    attribute = deepcopy(stix2misp_mapping.network_socket_extension_mapping[pattern_type])
+                    if pattern_type.startswith("network-traffic:extensions.'socket-ext'.is_"):
+                        if pattern_value != 'True':
+                            continue
+                        pattern_value = pattern_type.split('_')[1]
+                else:
+                    if pattern_type.startswith('network-traffic:protocols['):
+                        attributes.append({'type': 'text', 'object_relation': 'protocol', 'value': pattern_value})
+                    elif any(pattern_type.startswith(f'network-traffic:{feature}_ref') for feature in ('src', 'dst')):
+                        feature_type, ref = pattern_type.split(':')[1].split('_')
+                        ref, feature = ref.split('.')
+                        ref = f"{feature_type}_{'0' if ref == 'ref' else ref.strip('ref[]')}"
+                        references[ref].update(self._parse_network_connection_reference(feature_type, feature, pattern_value))
+                    continue
+            else:
+                attribute = deepcopy(stix2misp_mapping.network_traffic_mapping[pattern_type])
+            attribute['value'] = pattern_value
+            attributes.append(attribute)
+        attributes.extend(attribute for attribute in references.values())
+        return attributes
+
+    def parse_pe_pattern(self, pattern):
         attributes = []
         sections = defaultdict(dict)
-        pe = MISPObject('pe', misp_objects_path_custom=_MISP_objects_path)
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            p_value = p_value[1:-1]
-            if ':extensions.' in p_type:
-                if '.sections[' in p_type:
-                    p_type_list = p_type.split('.')
-                    stix_type = "hashes.{}".format(p_type_list[4][1:-1]) if '.hashes.' in p_type else p_type_list[3]
-                    sections[p_type_list[2]][stix_type] = p_value
+        pe = MISPObject('pe', misp_objects_path_custom=_misp_objects_path)
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            if ':extensions.' in pattern_type:
+                if '.sections[' in pattern_type:
+                    pattern_type = pattern_type.split('.')
+                    relation = pattern_type[-1].strip("'")
+                    if relation in stix2misp_mapping.pe_section_mapping:
+                        sections[pattern_type[2][-2]][relation] = pattern_value.strip("'")
                 else:
-                    stix_type = p_type.split('.')[-1]
-                    try:
-                        mapping = pe_mapping[stix_type]
-                        pe.add_attribute(**{'type': mapping['type'], 'object_relation': mapping['relation'],
-                                            'value': p_value, 'to_ids': True})
-                    except KeyError:
-                        if stix_type.startswith("x_misp_"):
-                            attribute_type, relation = self.parse_custom_property(stix_type)
-                            pe.add_attribute(**{'type': attribute_type, 'object_relation': relation[:-2],
-                                                'value': p_value, 'to_ids': False})
+                    pattern_type = pattern_type.split('.')[-1]
+                    if pattern_type not in stix2misp_mapping.pe_mapping:
+                        if pattern_type.startswith('x_misp_'):
+                            attribute = self.parse_custom_property(pattern_type)
+                            attribute['value'] = pattern_value.strip("'")
+                            pe.add_attribute(**attribute)
+                        continue
+                    attribute = deepcopy(stix2misp_mapping.pe_mapping[pattern_type])
+                    attribute['value'] = pattern_value.strip("'")
+                    pe.add_attribute(**attribute)
             else:
-                if 'file:hashes.' in p_type :
-                    _, h = p_type.split('.')
-                    h = h[1:-1]
-                    attributes.append({'type': h, 'object_relation': h, 'value': p_value, 'to_ids': True})
-                else:
-                    try:
-                        mapping = file_mapping[p_type]
-                        attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                           'value': p_value, 'to_ids': True})
-                    except KeyError:
-                        if "x_misp_" in  p_type:
-                            attribute_type, relation = p_type.split("x_misp_")[1][:-1].split("_")
-                            attributes.append({'type': attribute_type, 'object_relation': relation,
-                                               'value': p_value, 'to_ids': True})
-        for _, section in sections.items():
-            pe_section = MISPObject('pe-section', misp_objects_path_custom=_MISP_objects_path)
-            for stix_type, value in section.items():
-                if 'hashes.' in stix_type:
-                    h_type = stix_type.split('.')[1]
-                    pe_section.add_attribute(**{'type': h_type, 'object_relation': h_type,
-                                                'value': value, 'to_ids': True})
-                else:
-                    try:
-                        mapping = pe_section_mapping[stix_type]
-                        pe_section.add_attribute(**{'type': mapping['type'], 'object_relation': mapping['relation'],
-                                                    'value': value, 'to_ids': True})
-                    except KeyError:
-                        if "x_misp_" in  stix_type:
-                            attribute_type, relation = stix_type.split("x_misp_")[1][:-1].split("_")
-                            attributes.append({'type': attribute_type, 'object_relation': relation,
-                                               'value': value, 'to_ids': True})
+                if pattern_type not in stix2misp_mapping.file_mapping:
+                    if pattern_type.startswith('x_misp_'):
+                        attribute = self.parse_custom_property(pattern_type)
+                        attribute['value'] = pattern_value.strip("'")
+                        attributes.append(attribute)
+                    continue
+                attribute = deepcopy(stix2misp_mapping.file_mapping[pattern_type])
+                attribute['value'] = pattern_value.strip("'")
+                attributes.append(attribute)
+        for section in sections.values():
+            pe_section = MISPObject('pe-section', misp_objects_path_custom=_misp_objects_path)
+            for feature, value in section.items():
+                attribute = deepcopy(stix2misp_mapping.pe_section_mapping[feature])
+                attribute['value'] = value
+                pe_section.add_attribute(**attribute)
             pe.add_reference(pe_section.uuid, 'includes')
             self.misp_event.add_object(**pe_section)
         self.misp_event.add_object(**pe)
         return attributes, pe.uuid
 
-    def pattern_asn(self, pattern):
-        return self.fill_pattern_attributes(pattern, asn_mapping)
+    def parse_process_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'process_mapping')
 
-    def pattern_domain_ip(self, pattern):
-        return self.fill_pattern_attributes(pattern, domain_ip_mapping)
+    def parse_regkey_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'regkey_mapping')
 
-    def observable_ip_port(self, observable):
-        attributes, _ = self.attributes_from_network_traffic(observable)
-        return attributes
-
-    def pattern_ip_port(self, pattern):
-        return self.fill_pattern_attributes(pattern, network_traffic_mapping)
+    def parse_url_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'url_mapping')
 
     @staticmethod
-    def pattern_process(pattern):
+    def parse_user_account_pattern(pattern):
         attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            try:
-                mapping = process_mapping[p_type]
-            except KeyError:
+        for pattern_part in pattern:
+            pattern_type, pattern_value = pattern_part.split(' = ')
+            pattern_type = pattern_type.split('.')[-1].split('[')[0] if "extensions.'unix-account-ext'" in pattern_type else pattern_type.split(':')[-1]
+            if pattern_type not in stix2misp_mapping.user_account_mapping:
+                if pattern_type.startswith('group'):
+                    attributes.append({'type': 'text', 'object_relation': 'group', 'value': pattern_value.strip("'")})
                 continue
-            if p_type == 'process:child_refs':
-                for value in p_value[1:-1].split(','):
-                    attribute.append({'type': mapping['type'], 'value': value.strip(),
-                                     'object_relation': mapping['relation']})
-            else:
-                attributes.append({'type': mapping['type'], 'value': p_value,
-                                   'object_relation': mapping['relation']})
+            attribute = deepcopy(stix2misp_mapping.user_account_mapping[pattern_type])
+            attribute['value'] = pattern_value.strip("'")
+            attributes.append(attribute)
         return attributes
 
-    @staticmethod
-    def pattern_regkey(pattern):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            try:
-                mapping = regkey_mapping[p_type]
-            except KeyError:
-                continue
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': p_value.replace('\\\\', '\\')[1:-1]})
-        return attributes
-
-    def observable_socket(self, observable):
-        attributes, _ = self.attributes_from_network_traffic(observable)
-        return attributes
-
-    @staticmethod
-    def pattern_socket(pattern):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            p_value = p_value[1:-1]
-            try:
-                mapping = network_traffic_mapping[p_type]
-            except KeyError:
-                continue
-            if "network-traffic:extensions.'socket-ext'.is_" in p_type:
-                p_value = p_type.split('_')[1]
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': p_value})
-        return attributes
-
-    def pattern_url(self, pattern):
-        return self.fill_pattern_attributes(pattern, url_mapping)
-
-    def pattern_x509(self, pattern):
-        return self.fill_pattern_attributes(pattern, x509_mapping)
+    def parse_x509_pattern(self, pattern):
+        return self.fill_pattern_attributes(pattern, 'x509_mapping')
 
     ################################################################################
     ##                             UTILITY FUNCTIONS.                             ##
     ################################################################################
 
-    def fill_observable_attributes(self, stix_object, object_mapping):
-        attributes = []
-        for o_key, o_value in stix_object.items():
-            try:
-                mapping = object_mapping[o_key]
-                attributes.append({'type': mapping.get('type'), 'object_relation': mapping.get('relation'),
-                                   'value': o_value, 'to_ids': False})
-            except KeyError:
-                if "x_misp_" in o_key:
-                    attribute_type, relation = self.parse_custom_property(o_key)
-                    if isinstance(o_value, list):
-                        for v in o_value:
-                            attributes.append({'type': attribute_type, 'object_relation': relation[:-1],
-                                               'value': v, 'to_ids': False})
-                    else:
-                        attributes.append({'type': attribute_type, 'object_relation': relation[:-1],
-                                           'value': o_value, 'to_ids': False})
-        return attributes
+    def create_attribute_dict(self, stix_object):
+        labels = stix_object['labels']
+        attribute_uuid = stix_object.id.split('--')[1]
+        attribute = {'uuid': attribute_uuid,
+                     'type': self.get_misp_type(labels),
+                     'category': self.get_misp_category(labels)}
+        tags = [{'name': label} for label in labels[3:]]
+        if tags:
+            attribute['Tag'] = tags
+        attribute.update(self.parse_timeline(stix_object))
+        return attribute
 
-    def fill_object_attributes_observable(self, misp_object, mapping_dict, stix_object):
-        for stix_type, value in stix_object.items():
-            try:
-                mapping = mapping_dict[stix_type]
-                misp_object.add_attribute(**{'type': mapping['type'], 'object_relation': mapping['relation'],
-                                             'value': value, 'to_ids': False})
-            except KeyError:
-                if stix_type.startswith("x_misp_"):
-                    attribute_type, relation = self.parse_custom_property(stix_type)
-                    misp_object.add_attribute(**{'type': attribute_type, 'object_relation': relation[:-1],
-                                                 'value': value, 'to_ids': False})
-
-    def fill_pattern_attributes(self, pattern, object_mapping):
-        attributes = []
-        for p in pattern:
-            p_type, p_value = p.split(' = ')
-            try:
-                mapping = object_mapping[p_type]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'value': p_value[1:-1], 'to_ids': True})
-            except KeyError:
-                if "x_misp_" in p_type:
-                    attribute_type, relation = self.parse_custom_property(p_type)
-                    attributes.append({'type': attribute_type, 'object_relation': relation[:-2],
-                                       'value': p_value[1:-1], 'to_ids': True})
-        return attributes
+    def create_misp_object(self, stix_object):
+        labels = stix_object['labels']
+        object_type = self.get_misp_type(labels)
+        misp_object = MISPObject('file' if object_type == 'WindowsPEBinaryFile' else object_type,
+                                 misp_objects_path_custom=_misp_objects_path)
+        misp_object.uuid = stix_object.id.split('--')[1]
+        misp_object.update(self.parse_timeline(stix_object))
+        return misp_object, object_type
 
     @staticmethod
-    def parse_custom_property(p_type):
-        d_type = p_type.split("_")
-        attribute_type = d_type[2]
-        relation = "".join("{}-".format(t) for t in d_type[3:])
-        return attribute_type, relation
+    def _fill_object_attribute(feature, value):
+        return {'value': str(value) if feature in ('entropy', 'size') else value}
 
     @staticmethod
-    def parse_marking(marking):
-        marking_type = marking.definition_type
-        tag = getattr(marking.definition, marking_type)
-        return "{}:{}".format(marking_type, tag)
+    def _fill_observable_object_attribute(feature, value):
+        return {'value': str(value) if feature in ('entropy', 'size') else value,
+                'to_ids': False}
 
     @staticmethod
-    def parse_pattern(pattern):
+    def get_misp_category(labels):
+        return labels[1].split('=')[1].strip('"')
+
+    @staticmethod
+    def get_misp_type(labels):
+        return labels[0].split('=')[1].strip('"')
+
+    @staticmethod
+    def parse_attribute_pattern(pattern):
         if ' AND ' in pattern:
-            pattern_parts = pattern.split(' AND ')
+            pattern_parts = pattern.strip('[]').split(' AND ')
             if len(pattern_parts) == 3:
                 _, value1 = pattern_parts[2].split(' = ')
                 _, value2 = pattern_parts[0].split(' = ')
-                return '{}|{}'.format(value1[1:-2], value2[1:-1])
+                return '{}|{}'.format(value1.strip("'"), value2.strip("'"))
             else:
                 _, value1 = pattern_parts[0].split(' = ')
                 _, value2 = pattern_parts[1].split(' = ')
                 if value1 in ("'ipv4-addr'", "'ipv6-addr'"):
-                    return value2[1:-2]
-                return '{}|{}'.format(value1[1:-1], value2[1:-2])
+                    return value2.strip("'")
+                return '{}|{}'.format(value1.strip("'"), value2.strip("'"))
         else:
-            return pattern.split(' = ')[1][1:-2]
+            return pattern.split(' = ')[1].strip("]'")
 
-    def parse_pattern_with_data(self, pattern):
-        if 'artifact:payload_bin' not in pattern:
-            return self.parse_pattern(pattern)
-        pattern_parts = pattern.split(' AND ')
+    def parse_attribute_pattern_with_data(self, pattern):
+        if 'file:content_ref.payload_bin' not in pattern:
+            return self.parse_attribute_pattern(pattern)
+        pattern_parts = pattern.strip('[]').split(' AND ')
         if len(pattern_parts) == 3:
             filename = pattern_parts[0].split(' = ')[1]
             md5 = pattern_parts[1].split(' = ')[1]
-            return "{}|{}".format(filename[1:-1], md5[1:-1]), pattern_parts[2].split(' = ')[1][1:-2]
-        return pattern_parts[0].split(' = ')[1][1:-1], pattern_parts[1].split(' = ')[1][1:-2]
+            return "{}|{}".format(filename.strip("'"), md5.strip("'")), pattern_parts[2].split(' = ')[1].strip("'")
+        return pattern_parts[0].split(' = ')[1].strip("'"), pattern_parts[1].split(' = ')[1].strip("'")
+
+    @staticmethod
+    def parse_custom_property(property):
+        properties = property.split('_')
+        return {'type': properties[2], 'object_relation': '-'.join(properties[3:])}
 
 
 class ExternalStixParser(StixParser):
+    _object_from_refs = {'course-of-action': 'parse_course_of_action', 'vulnerability': 'parse_external_vulnerability',
+                          'indicator': 'parse_external_indicator', 'observed-data': 'parse_external_observable'}
+    _observable_mapping = {('artifact', 'file'): 'parse_file_observable',
+                           ('artifact', 'directory', 'file'): 'parse_file_observable',
+                           ('artifact', 'email-addr', 'email-message', 'file'): 'parse_email_observable',
+                           ('autonomous-system',): 'parse_asn_observable',
+                           ('autonomous-system', 'ipv4-addr'): 'parse_asn_observable',
+                           ('autonomous-system', 'ipv6-addr'): 'parse_asn_observable',
+                           ('autonomous-system', 'ipv4-addr', 'ipv6-addr'): 'parse_asn_observable',
+                           ('directory', 'file'): 'parse_file_observable',
+                           ('domain-name',): 'parse_domain_ip_observable',
+                           ('domain-name', 'ipv4-addr'): 'parse_domain_ip_observable',
+                           ('domain-name', 'ipv6-addr'): 'parse_domain_ip_observable',
+                           ('domain-name', 'ipv4-addr', 'ipv6-addr'): 'parse_domain_ip_observable',
+                           ('domain-name', 'ipv4-addr', 'network-traffic'): 'parse_domain_ip_network_traffic_observable',
+                           ('domain-name', 'ipv6-addr', 'network-traffic'): 'parse_domain_ip_network_traffic_observable',
+                           ('domain-name', 'ipv4-addr', 'ipv6-addr', 'network-traffic'): 'parse_domain_ip_network_traffic_observable',
+                           ('domain-name', 'network-traffic'): 'parse_domain_network_traffic_observable',
+                           ('domain-name', 'network-traffic', 'url'): 'parse_url_observable',
+                           ('email-addr',): 'parse_email_address_observable',
+                           ('email-addr', 'email-message'): 'parse_email_observable',
+                           ('email-addr', 'email-message', 'file'): 'parse_email_observable',
+                           ('email-message',): 'parse_email_observable',
+                           ('file',): 'parse_file_observable',
+                           ('file', 'process'): 'parse_process_observable',
+                           ('ipv4-addr',): 'parse_ip_address_observable',
+                           ('ipv6-addr',): 'parse_ip_address_observable',
+                           ('ipv4-addr', 'network-traffic'): 'parse_ip_network_traffic_observable',
+                           ('ipv6-addr', 'network-traffic'): 'parse_ip_network_traffic_observable',
+                           ('ipv4-addr', 'ipv6-addr', 'network-traffic'): 'parse_ip_network_traffic_observable',
+                           ('mac-addr',): 'parse_mac_address_observable',
+                           ('mutex',): 'parse_mutex_observable',
+                           ('process',): 'parse_process_observable',
+                           ('x509-certificate',): 'parse_x509_observable',
+                           ('url',): 'parse_url_observable',
+                           ('user-account',): 'parse_user_account_observable',
+                           ('windows-registry-key',): 'parse_regkey_observable'}
+    _pattern_mapping = {('artifact', 'file'): 'parse_file_pattern',
+                        ('artifact', 'directory', 'file'): 'parse_file_pattern',
+                        ('autonomous-system', ): 'parse_as_pattern',
+                        ('autonomous-system', 'ipv4-addr'): 'parse_as_pattern',
+                        ('autonomous-system', 'ipv6-addr'): 'parse_as_pattern',
+                        ('autonomous-system', 'ipv4-addr', 'ipv6-addr'): 'parse_as_pattern',
+                        ('directory',): 'parse_file_pattern',
+                        ('directory', 'file'): 'parse_file_pattern',
+                        ('domain-name',): 'parse_domain_ip_port_pattern',
+                        ('domain-name', 'ipv4-addr', 'url'): 'parse_domain_ip_port_pattern',
+                        ('domain-name', 'ipv6-addr', 'url'): 'parse_domain_ip_port_pattern',
+                        ('domain-name', 'network-traffic'): 'parse_domain_ip_port_pattern',
+                        ('domain-name', 'network-traffic', 'url'): 'parse_url_pattern',
+                        ('email-addr',): 'parse_email_address_pattern',
+                        ('email-message',): 'parse_email_message_pattern',
+                        ('file',): 'parse_file_pattern',
+                        ('ipv4-addr',): 'parse_ip_address_pattern',
+                        ('ipv6-addr',): 'parse_ip_address_pattern',
+                        ('ipv4-addr', 'ipv6-addr'): 'parse_ip_address_pattern',
+                        ('mac-addr',): 'parse_mac_address_pattern',
+                        ('mutex',): 'parse_mutex_pattern',
+                        ('network-traffic',): 'parse_network_traffic_pattern',
+                        ('process',): 'parse_process_pattern',
+                        ('url',): 'parse_url_pattern',
+                        ('user-account',): 'parse_user_account_pattern',
+                        ('windows-registry-key',): 'parse_regkey_pattern',
+                        ('x509-certificate',): 'parse_x509_pattern'}
+    _pattern_forbidden_relations = (' LIKE ', ' FOLLOWEDBY ', ' MATCHES ', ' ISSUBSET ', ' ISSUPERSET ', ' REPEATS ')
+    _single_attribute_fields = ('type', 'value', 'to_ids')
+
     def __init__(self):
-        super(ExternalStixParser, self).__init__()
-        self.object_from_refs = {'course-of-action': 'parse_course_of_action', 'vulnerability': 'parse_external_vulnerability',
-                                 'indicator': 'parse_external_indicator', 'observed-data': 'parse_external_observable'}
-        self.observable_mapping = {('artifact', 'file'): 'parse_file_object_observable',
-                                   ('autonomous-system',): 'parse_asn_observable',
-                                   ('autonomous-system', 'ipv4-addr'): 'parse_asn_observable',
-                                   ('autonomous-system', 'ipv6-addr'): 'parse_asn_observable',
-                                   ('autonomous-system', 'ipv4-addr', 'ipv6-addr'): 'parse_asn_observable',
-                                   ('domain-name',): 'parse_domain_ip_observable',
-                                   ('domain-name', 'ipv4-addr'): 'parse_domain_ip_observable',
-                                   ('domain-name', 'ipv6-addr'): 'parse_domain_ip_observable',
-                                   ('domain-name', 'ipv4-addr', 'network-traffic'): 'parse_ip_port_or_network_socket_observable',
-                                   ('domain-name', 'ipv6-addr', 'network-traffic'): 'parse_ip_port_or_network_socket_observable',
-                                   ('domain-name', 'ipv4-addr', 'ipv6-addr', 'network-traffic'): 'parse_ip_port_or_network_socket_observable',
-                                   ('domain-name', 'network-traffic'): 'parse_network_socket_observable',
-                                   ('domain-name', 'network-traffic', 'url'): 'parse_url_object_observable',
-                                   ('email-addr',): 'parse_email_address_observable',
-                                   ('email-addr', 'email-message'): 'parse_email_observable',
-                                   ('email-addr', 'email-message', 'file'): 'parse_email_observable',
-                                   ('email-message',): 'parse_email_observable',
-                                   ('file',): 'parse_file_observable',
-                                   ('ipv4-addr',): 'parse_ip_address_observable',
-                                   ('ipv6-addr',): 'parse_ip_address_observable',
-                                   ('ipv4-addr', 'network-traffic'): 'parse_ip_network_traffic_observable',
-                                   ('ipv6-addr', 'network-traffic'): 'parse_ip_network_traffic_observable',
-                                   ('mac-addr',): 'parse_mac_address_observable',
-                                   ('mutex',): 'parse_mutex_observable',
-                                   ('process',): 'parse_process_observable',
-                                   ('x509-certificate',): 'parse_x509_observable',
-                                   ('url',): 'parse_url_observable',
-                                   ('user-account',): 'parse_user_account_observable',
-                                   ('windows-registry-key',): 'parse_regkey_observable'}
-        self.pattern_mapping = {('directory',): 'parse_file_pattern',
-                                ('directory', 'file'): 'parse_file_pattern',
-                                ('domain-name',): 'parse_domain_ip_port_pattern',
-                                ('domain-name', 'ipv4-addr', 'url'): 'parse_domain_ip_port_pattern',
-                                ('domain-name', 'ipv6-addr', 'url'): 'parse_domain_ip_port_pattern',
-                                ('email-addr',): 'parse_email_address_pattern',
-                                ('file',): 'parse_file_pattern',
-                                ('ipv4-addr',): 'parse_ip_address_pattern',
-                                ('ipv6-addr',): 'parse_ip_address_pattern',
-                                ('network-traffic',): 'parse_network_traffic_pattern',
-                                ('process',): 'parse_process_pattern',
-                                ('url',): 'parse_url_pattern',
-                                ('user-account',): 'parse_user_account_pattern',
-                                ('windows-registry-key',): 'parse_regkey_pattern',
-                                ('x509-certificate',): 'parse_x509_pattern'}
-        self.pattern_forbidden_relations = (' LIKE ', ' FOLLOWEDBY ', ' MATCHES ', ' ISSUBSET ', ' ISSUPERSET ', ' REPEATS ')
-        self.single_attribute_fields = ('type', 'value', 'to_ids')
-
-    def handler(self):
-        with open(self.synonyms_to_tag_names, 'rt', encoding='utf-8') as f:
-            synonyms_to_tag_names = json.loads(f.read())
-        self.synonyms_to_tag_names = synonyms_to_tag_names
-        self.version_attribute = {'type': 'text', 'object_relation': 'version', 'value': self.stix_version}
-        self.general_handler()
-
-    def parsing_process(self, object2parse, object_type):
-        try:
-            getattr(self, self.object_from_refs[object_type])(object2parse)
-        except KeyError:
-            print("Unknown {} type: {}".format(self.stix_version, object_type), file=sys.stderr)
-
-    def parse_galaxy(self, galaxy):
-        if galaxy.name in self.synonyms_to_tag_names:
-            return self.synonyms_to_tag_names[galaxy.name]
-        return ['misp-galaxy:{}="{}"'.format(galaxy._type, galaxy.name)]
-
-    def parse_external_indicator(self, indicator):
-        pattern = indicator.pattern
-        indicator_id = indicator.id.split('--')[1]
-        try:
-            if hasattr(indicator, 'object_marking_refs'):
-                self.parse_external_pattern(pattern, indicator_id, marking=indicator.object_marking_refs)
-            else:
-                self.parse_external_pattern(pattern, indicator_id)
-            # Deeper analyse of patterns coming when we get examples
-        except Exception:
-            self.add_stix2_pattern_object(pattern, indicator_id)
-
-    def parse_external_observable(self, observable):
-        objects = observable.objects
-        types = self.parse_external_observable_object(objects)
-        try:
-            to_call = self.observable_mapping[types]
-            observable_id = observable.id.split('--')[1]
-            if hasattr(observable, 'object_marking_refs'):
-                getattr(self, to_call)(objects, observable_id, marking=observable.object_marking_refs)
-            else:
-                getattr(self, to_call)(objects, observable_id)
-        except KeyError:
-            print('{} not parsed at the moment'.format(types), file=sys.stderr)
-        # deeper analyse to come, as well as for indicators
-
-    @staticmethod
-    def parse_external_observable_object(observable_objects):
-        types = set()
-        for _object in observable_objects.values():
-            types.add(_object._type)
-        return tuple(sorted(types))
-
-    def parse_external_pattern(self, pattern, uuid, marking=None):
-        if not any(relation in pattern for relation in self.pattern_forbidden_relations):
-            pattern = pattern[1:-1]
-            if ' OR ' in pattern and ' AND ' in pattern:
-                return
-            if ' OR ' in pattern:
-                pattern = pattern.split(' OR ')
-                for p in pattern:
-                    type_ = tuple([p.split(' = ')[0].split(':')[0]])
-                    try:
-                        getattr(self, self.pattern_mapping[type_])([p.strip()], marking)
-                    except KeyError:
-                        print('{} not parsed at the moment'.format(type_), file=sys.stderr)
-                        raise Exception
-            else:
-                pattern = [p.strip() for p in pattern.split(' AND ')]
-                types = self.parse_external_pattern_types(pattern)
-                try:
-                    getattr(self, self.pattern_mapping[types])(pattern, marking, uuid=uuid)
-                except KeyError:
-                    print('{} not parsed at the moment'.format(types), file=sys.stderr)
-                    raise Exception
-        else:
-            self.add_stix2_pattern_object(pattern, uuid)
-
-    @staticmethod
-    def parse_external_pattern_types(pattern):
-        types = set()
-        for p in pattern:
-            types.add(p.split('=')[0].split(':')[0])
-        return tuple(sorted(types))
-
-    def parse_external_vulnerability(self, o):
-        attribute = {'type': 'vulnerability', 'value': o.get('name')}
-        if 'description' in o:
-            attribute['comment'] = o.get('description')
-        if 'object_marking_refs' in o:
-            attribute = self.add_tag_in_attribute(attribute, o['object_marking_refs'])
-        self.misp_event.add_attribute(**attribute)
-
-    @staticmethod
-    def attributes_from_external_pattern(pattern):
-        types = self.parse_external_pattern_types()
-        pattern_type, pattern_value = pattern.split(' = ')
-        pattern_type, pattern_value = pattern_type[1:].strip(), pattern_value[1:-2].strip()
-        stix_type, value_type = pattern_type.split(':')
-        if 'hashes' in value_type and 'x509' not in stix_type:
-            h_type = value_type.split('.')[1].replace("'", '')
-            return {'type': h_type, 'value': pattern_value}
-        # Might cause some issues, need more examples to test
-        return {'type': external_pattern_mapping[stix_type][value_type].get('type'), 'value': pattern_value}
+        super().__init__()
+        self._stix2misp_mapping.update({'attack-pattern': 'parse_attack_pattern',
+                                        'course-of-action': 'parse_course_of_action',
+                                        'vulnerability': 'parse_vulnerability'})
 
     ################################################################################
     ##                             PARSING FUNCTIONS.                             ##
     ################################################################################
 
-    def add_attributes_from_observable(self, objects, attribute_type, identifier, uuid, marking):
-        attribute = {'to_ids': False}
-        if len(objects) == 1:
-            attribute['uuid'] = uuid
-        if marking:
-            attribute['type'] = attribute_type
-            for observable in objects.values():
-                attribute['value'] = getattr(observable, identifier)
-                attribute = self.add_tag_in_attribute(attribute, marking)
-                self.misp_event.add_attribute(**attribute)
-        else:
-            for observable in objects.values():
-                self.misp_event.add_attribute(attribute_type, getattr(observable, identifier), **attribute)
-
-    def add_attributes_from_pattern(self, attribute_type, pattern, marking, uuid):
-        _, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attribute = {'to_ids': True}
-        if len(pattern_values) == 1 and uuid is not None:
-            attribute['uuid'] = uuid
-            attribute['value'] = pattern_values[0]
-            attribute['type'] = attribute_type
-            if marking:
-                attribute = self.add_tag_in_attribute(attribute, marking)
-            self.handle_single_attribute(attribute, uuid=uuid)
-        else:
-            if marking:
-                attribute['type'] = attribute_type
-                for value in pattern_values:
-                    attribute['value'] = value
-                    attribute = self.add_tag_in_attribute(attribute, marking)
-                    self.misp_event.add_attribute(**attribute)
+    def parse_event(self, stix_objects):
+        for stix_object in stix_objects:
+            object_type = stix_object['type']
+            if object_type in self._stix2misp_mapping:
+                getattr(self, self._stix2misp_mapping[object_type])(stix_object)
             else:
-                for value in pattern_values:
-                    self.misp_event.add_attribute(attribute_type, value, **attribute)
+                print(f'not found: {object_type}', file=sys.stderr)
+        if self.relationship:
+            self.parse_relationships()
+        if self.galaxy:
+            self.parse_galaxies()
+        if hasattr(self, 'report'):
+            self.parse_report()
+        else:
+            self.misp_event.info = 'Imported with the STIX to MISP import script.'
+        if self.tags:
+            for tag in self.tags:
+                self.misp_event.add_tag(tag)
+
+    def parse_galaxy(self, galaxy):
+        if galaxy.name in self._synonyms_to_tag_names:
+            return self._synonyms_to_tag_names[galaxy.name]
+        return [f'misp-galaxy:{galaxy._type}="{galaxy.name}"']
+
+    def _parse_indicator(self, indicator):
+        pattern = indicator.pattern
+        if any(relation in pattern for relation in self._pattern_forbidden_relations) or all(relation in pattern for relation in (' OR ', ' AND ')):
+            self.add_stix2_pattern_object(indicator)
+        elif ' OR ' in pattern:
+            self.parse_multiple_attributes(indicator)
+        else:
+            self.parse_usual_indicator(indicator)
+
+    def _parse_observable(self, observable):
+        types = self._parse_observable_types(observable.objects)
+        print(f'has marking refs: {hasattr(observable, "object_marking_refs")}')
+        try:
+            getattr(self, self._observable_mapping[types])(observable)
+        except KeyError:
+            print(f'Type(s) not supported at the moment: {types}\n', file=sys.stderr)
+
+    def _parse_undefined(self, stix_object):
+        try:
+            self.objects_to_parse[stix_object['id'].split('--')[1]] = stix_object
+        except AttributeError:
+            self.objects_to_parse = {stix_object['id'].split('--')[1]: stix_object}
+
+    def add_stix2_pattern_object(self, indicator):
+        misp_object = MISPObject('stix2-pattern', misp_objects_path_custom=_misp_objects_path)
+        misp_object.uuid = indicator.id.split('--')[1]
+        misp_object.update(self.parse_timeline(indicator))
+        version = f'STIX {indicator.pattern_version}' if hasattr(indicator, 'pattern_version') else 'STIX 2.0'
+        misp_object.add_attribute(**{'type': 'text', 'object_relation': 'version', 'value': version})
+        misp_object.add_attribute(**{'type': 'stix2-pattern', 'object_relation': 'stix2-pattern',
+                                     'value': indicator.pattern})
+        self.misp_event.add_object(**misp_object)
 
     @staticmethod
-    def  attributes_from_dict(values, mapping_dict, to_ids):
+    def fill_misp_object(misp_object, stix_object, mapping):
+        for key, feature in getattr(stix2misp_mapping, mapping).items():
+            if hasattr(stix_object, key):
+                attribute = deepcopy(feature)
+                attribute['value'] = getattr(stix_object, key)
+                misp_object.add_attribute(**attribute)
+
+    @staticmethod
+    def fill_misp_object_from_dict(misp_object, stix_object, mapping):
+        for key, feature in getattr(stix2misp_mapping, mapping).items():
+            if key in stix_object:
+                attribute = deepcopy(feature)
+                attribute['value'] = stix_object[key]
+                misp_object.add_attribute(**attribute)
+
+    def parse_attack_pattern(self, attack_pattern):
+        misp_object = self.create_misp_object(attack_pattern)
+        if hasattr(attack_pattern, 'external_references'):
+            for reference in attack_pattern.external_references:
+                source_name = reference['source_name']
+                value = reference['external_id'].split('-')[1] if source_name == 'capec' else reference['url']
+                attribute = deepcopy(stix2misp_mapping.attack_pattern_references_mapping[source_name]) if source_name in stix2misp_mapping.attack_pattern_references_mapping else stix2misp_mapping.references_attribute_mapping
+                attribute['value'] = value
+                misp_object.add_attribute(**attribute)
+        self.fill_misp_object(misp_object, attack_pattern, 'attack_pattern_mapping')
+        self.misp_event.add_object(**misp_object)
+
+    def parse_course_of_action(self, course_of_action):
+        misp_object = self.create_misp_object(course_of_action)
+        self.fill_misp_object(misp_object, course_of_action, 'course_of_action_mapping')
+        self.misp_event.add_object(**misp_object)
+
+    def parse_usual_indicator(self, indicator):
+        pattern = tuple(part.strip() for part in indicator.pattern.strip('[]').split(' AND '))
+        types = self._parse_pattern_types(pattern)
+        try:
+            getattr(self, self._pattern_mapping[types])(indicator)
+        except KeyError:
+            print(f'Type(s) not supported at the moment: {types}\n', file=sys.stderr)
+            self.add_stix2_pattern_object(indicator)
+
+    def parse_vulnerability(self, vulnerability):
+        attributes = self._get_attributes_from_observable(vulnerability, 'vulnerability_mapping')
+        if hasattr(vulnerability, 'external_references'):
+            for reference in vulnerability.external_references:
+                if reference['source_name'] == 'url':
+                    attribute = deepcopy(stix2misp_mapping.references_attribute_mapping)
+                    attribute['value'] = reference['url']
+                    attributes.append(attribute)
+        if len(attributes) == 1 and attributes[0]['object_relation'] == 'id':
+            attributes[0]['type'] = 'vulnerability'
+        self.handle_import_case(vulnerability, attributes, 'vulnerability')
+
+    ################################################################################
+    ##                        OBSERVABLE PARSING FUNCTIONS                        ##
+    ################################################################################
+
+    @staticmethod
+    def _fetch_reference_type(references, object_type):
+        for key, reference in references.items():
+            if isinstance(reference, getattr(stix2, object_type)):
+                return key
+        return None
+
+    @staticmethod
+    def _fetch_user_account_type_observable(observable_objects):
+        for observable_object in observable_objects.values():
+            if hasattr(observable_object, 'extensions') or any(key not in ('user_id', 'credential', 'type') for key in observable_object):
+                return 'user-account', 'user_account_mapping'
+        return 'credential', 'credential_mapping'
+
+    @staticmethod
+    def _get_attributes_from_observable(stix_object, mapping):
         attributes = []
-        for type_, value in values.items():
-            try:
-                mapping = mapping_dict[type_]
-            except KeyError:
-                continue
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': value, 'to_ids': to_ids})
+        for key, value in stix_object.items():
+            if key in getattr(stix2misp_mapping, mapping) and value:
+                attribute = deepcopy(getattr(stix2misp_mapping, mapping)[key])
+                attribute.update({'value': value, 'to_ids': False})
+                attributes.append(attribute)
         return attributes
 
-    def handle_pe_case(self, extension, attributes, uuid):
+    def get_network_traffic_attributes(self, network_traffic, references):
+        attributes = self._get_attributes_from_observable(network_traffic, 'network_traffic_mapping')
+        mapping = 'network_traffic_references_mapping'
+        attributes.extend(self.parse_network_traffic_references(network_traffic, references, mapping))
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, mapping, 'dst'))
+        return attributes
+
+    @staticmethod
+    def _handle_attachment_type(stix_object, is_reference, filename):
+        _has_md5 = hasattr(stix_object, 'hashes') and 'MD5' in stix_object.hashes
+        if is_reference and _has_md5:
+            return 'malware-sample', f'{filename}|{stix_object.hashes["MD5"]}'
+        return 'attachment', filename
+
+    def handle_pe_observable(self, attributes, extension, observable):
         pe_uuid = self.parse_pe(extension)
-        file_object = self.create_misp_object(attributes, 'file', uuid)
-        file_object.add_reference(pe_uuid, 'includes')
+        file = self.create_misp_object(observable, 'file')
+        file.add_reference(pe_uuid, 'includes')
+        for attribute in attributes:
+            file.add_attribute(**attribute)
+        self.misp_event.add_object(**file)
+
+    @staticmethod
+    def _is_reference(network_traffic, reference):
+        for feature in ('src', 'dst'):
+            for reference_type in (f'{feature}_{ref}' for ref in ('ref', 'refs')):
+                if reference in network_traffic.get(reference_type, []):
+                    return True
+        return False
+
+    @staticmethod
+    def _network_traffic_has_extension(network_traffic):
+        if not hasattr(network_traffic, 'extensions'):
+            return None
+        if 'socket-ext' in network_traffic.extensions:
+            return 'parse_socket_extension_observable'
+        return None
+
+    def parse_asn_observable(self, observable):
+        autonomous_system, references = self.filter_main_object(observable.objects, 'AutonomousSystem')
+        mapping = 'asn_mapping'
+        attributes = self._get_attributes_from_observable(autonomous_system, mapping)
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, mapping))
+        self.handle_import_case(observable, attributes, 'asn')
+
+    def parse_domain_ip_observable(self, observable):
+        domain, references = self.filter_main_object(observable.objects, 'DomainName')
+        mapping = 'domain_ip_mapping'
+        attributes = [self._parse_observable_reference(domain, mapping)]
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, mapping))
+        self.handle_import_case(observable, attributes, 'domain-ip')
+
+    def parse_domain_ip_network_traffic_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable.objects, 'NetworkTraffic')
+        extension = self._network_traffic_has_extension(network_traffic)
+        if extension:
+            attributes, object_name = getattr(self, extension)(network_traffic, references)
+            return self.handle_import_case(observable, attributes, object_name)
+        if self._required_protocols(network_traffic.protocols):
+            attributes = self.parse_network_connection_object(network_traffic, references)
+            return self.handle_import_case(observable, attributes, 'network-connection')
+        attributes, object_name = self.parse_network_traffic_objects(network_traffic, references)
+        self.handle_import_case(observable, attributes, object_name)
+
+    def parse_domain_network_traffic_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable.objects, 'NetworkTraffic')
+        extension = self._network_traffic_has_extension(network_traffic)
+        if extension:
+            attributes, object_name = getattr(self, extension)(network_traffic, references)
+            return self.handle_import_case(observable, attributes, object_name)
+        attributes = self.parse_network_connection_object(network_traffic, references)
+        self.handle_import_case(observable, attributes, 'network-connection')
+
+    def parse_email_observable(self, observable):
+        email_message, references = self.filter_main_object(observable.objects, 'EmailMessage')
+        attributes = self._get_attributes_from_observable(email_message, 'email_mapping')
+        if hasattr(email_message, 'additional_header_fields'):
+            attributes.extend(self._get_attributes_from_observable(email_message.additional_header_fields, 'email_mapping'))
+        attributes.extend(self._parse_email_references(email_message, references))
+        if hasattr(email_message, 'body_multipart') and email_message.body_multipart:
+            attributes.extend(self._parse_email_body(email_message.body_multipart, references))
+        if references:
+            print(f'Unable to parse the following observable objects: {references}', file=sys.stderr)
+        self.handle_import_case(observable, attributes, 'email')
+
+    def parse_file_observable(self, observable):
+        file, references = self.filter_main_object(observable.objects, 'File')
+        attributes = self._get_attributes_from_observable(file, 'file_mapping')
+        if 'hashes' in file:
+            attributes.extend(self._get_attributes_from_observable(file.hashes, 'file_mapping'))
+        if references:
+            filename = file.name if hasattr(file, 'name') else 'unknown_filename'
+            for key, reference in references.items():
+                if isinstance(reference, stix2.Artifact):
+                    _is_content_ref = 'content_ref' in file and file.content_ref == key
+                    attribute_type, value = self._handle_attachment_type(reference, _is_content_ref, filename)
+                    attribute = {
+                        'type': attribute_type,
+                        'object_relation': attribute_type,
+                        'value': value,
+                        'to_ids': False
+                    }
+                    if hasattr(reference, 'payload_bin'):
+                        attribute['data'] = reference.payload_bin
+                    attributes.append(attribute)
+                elif isinstance(reference, stix2.Directory):
+                    attribute = {
+                        'type': 'text',
+                        'object_relation': 'path',
+                        'value': reference.path,
+                        'to_ids': False
+                    }
+                    attributes.append(attribute)
+        if hasattr(file, 'extensions'):
+            # Support of more extension types probably in the future
+            if 'windows-pebinary-ext' in file.extensions:
+                # Here we do not go to the standard route of "handle_import_case"
+                # because we want to make sure a file object is created
+                return self.handle_pe_observable(attributes, file.extensions['windows-pebinary-ext'], observable)
+            extension_types = (extension_type for extension_type in file.extensions.keys())
+            print(f'File extension type(s) not supported at the moment: {", ".join(extension_types)}', file=sys.stderr)
+        self.handle_import_case(observable, attributes, 'file')
+
+    def parse_ip_network_traffic_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable.objects, 'NetworkTraffic')
+        extension = self._network_traffic_has_extension(network_traffic)
+        if extension:
+            attributes, object_name = getattr(self, extension)(network_traffic, references)
+            return self.handle_import_case(observable, attributes, object_name)
+        attributes = self.parse_ip_port_object(network_traffic, references)
+        self.handle_import_case(observable, attributes, 'ip-port')
+
+    def parse_ip_port_object(self, network_traffic, references):
+        attributes = self._get_attributes_from_observable(network_traffic, 'network_traffic_mapping')
+        attributes.extend(self.parse_network_traffic_references(network_traffic, references, 'ip_port_references_mapping'))
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, 'domain_ip_mapping'))
+        return attributes
+
+    def parse_mac_address_observable(self, observable):
+        args = ('mac-address', 'value')
+        if len(observable.objects) == 1:
+            self.add_attribute_from_observable(observable, *args)
+        else:
+            self.add_attributes_from_observable(observable.objects, *args)
+
+    def parse_network_connection_object(self, network_traffic, references):
+        attributes = self.get_network_traffic_attributes(network_traffic, references)
+        attributes.extend(self.parse_protocols(network_traffic.protocols, 'observable object'))
+        return attributes
+
+    def parse_network_traffic_objects(self, network_traffic, references):
+        _has_domain = self._fetch_reference_type(references.values(), 'DomainName')
+        if _has_domain and _is_reference(network_traffic, _has_domain):
+            return self.parse_network_connection_object(network_traffic, references), 'network-connection'
+        return self.parse_ip_port_object(network_traffic, references), 'ip-port'
+
+    def parse_network_traffic_references(self, network_traffic, references, mapping):
+        attributes = []
+        for feature in ('src', 'dst'):
+            ref = f'{feature}_ref'
+            if hasattr(network_traffic, ref):
+                reference = getattr(network_traffic, ref)
+                attributes.append(self._parse_observable_reference(references.pop(reference), mapping, feature))
+            if hasattr(network_traffic, f'{ref}s'):
+                for reference in getattr(network_traffic, f'{ref}s'):
+                    attributes.append(self._parse_observable_reference(references.pop(reference), mapping, feature))
+        return attributes
+
+    def parse_mutex_observable(self, observable):
+        args = ('mutex', 'name')
+        if len(observable.objects) == 1:
+            self.add_attribute_from_observable(observable, *args)
+        else:
+            self.add_attributes_from_observable(observable.objects, *args)
+
+    def parse_process_observable(self, observable):
+        process, references = self.filter_main_object(observable.objects, 'Process', test_function='_process_test_filter')
+        attributes = self._get_attributes_from_observable(process, 'process_mapping')
+        if hasattr(process, 'parent_ref'):
+            attributes.extend(self._get_attributes_from_observable(references.pop(process.parent_ref), 'parent_process_reference_mapping'))
+        if hasattr(process, 'child_refs'):
+            for reference in process.child_refs:
+                attributes.extend(self._get_attributes_from_observable(references.pop(reference), 'child_process_reference_mapping'))
+        if hasattr(process, 'binary_ref'):
+            reference = references.pop(process.binary_ref)
+            attribute = {
+                'value': reference.name,
+                'to_ids': False
+            }
+            attribute.update(stix2misp_mapping.process_image_mapping)
+            attributes.append(attribute)
+        if references:
+            print(f'Unable to parse the following observable objects: {references}', file=sys.stderr)
+        self.handle_import_case(observable, attributes, 'process')
+
+    def parse_protocols(self, protocols, object_type):
+        attributes = []
+        protocols = (protocol.upper() for protocol in protocols)
+        for protocol in protocols:
+            try:
+                attributes.append(self._parse_network_traffic_protocol(protocol))
+            except KeyError:
+                print(f'Unknown protocol in network-traffic {object_type}: {protocol}', file=sys.stderr)
+        return attributes
+
+    def parse_regkey_observable(self, observable):
+        attributes = []
+        for observable_object in observable.objects.values():
+            attributes.extend(self._get_attributes_from_observable(observable_object, 'regkey_mapping'))
+            if 'values' in observable_object:
+                for registry_value in observable_object['values']:
+                    attributes.extend(self._get_attributes_from_observable(registry_value, 'regkey_mapping'))
+        self.handle_import_case(observable, attributes, 'registry-key')
+
+    def parse_socket_extension_observable(self, network_traffic, references):
+        attributes = self.get_network_traffic_attributes(network_traffic, references)
+        for key, value in network_traffic.extensions['socket-ext'].items():
+            if key not in stix2misp_mapping.network_socket_extension_mapping:
+                print(f'Unknown socket extension field in observable object: {key}', file=sys.stderr)
+                continue
+            if key.startswith('is_') and not value:
+                continue
+            attribute = {
+                'value': key.split('_')[1] if key.startswith('is_') else value,
+                'to_ids': False
+            }
+            attribute.update(stix2misp_mapping.network_socket_extension_mapping[key])
+            attributes.append(attribute)
+        return attributes, 'network-socket'
+
+    def parse_url_observable(self, observable):
+        network_traffic, references = self.filter_main_object(observable.objects, 'NetworkTraffic')
+        attributes = self._get_attributes_from_observable(network_traffic, 'network_traffic_mapping') if network_traffic else []
+        if references:
+            for reference in references.values():
+                attributes.append(self._parse_observable_reference(reference, 'url_mapping'))
+        self.handle_import_case(observable, attributes, 'url')
+
+    def parse_user_account_extension(self, extension):
+        attributes = self._parse_user_account_groups(extension['groups']) if 'groups' in extension else []
+        attributes.extend(self._get_attributes_from_observable(extension, 'user_account_mapping'))
+        return attributes
+
+    def parse_user_account_observable(self, observable):
+        attributes = []
+        object_name, mapping = self._fetch_user_account_type_observable(observable.objects)
+        for observable_object in observable.objects.values():
+            attributes.extend(self._get_attributes_from_observable(observable_object, mapping))
+            if hasattr(observable_object, 'extensions') and observable_object.extensions.get('unix-account-ext'):
+                attributes.extend(self.parse_user_account_extension(observable_object.extensions['unix-account-ext']))
+        self.handle_import_case(observable, attributes, object_name)
+
+    def parse_x509_observable(self, observable):
+        attributes = []
+        for observable_object in observable.objects.values():
+            attributes.extend(self._get_attributes_from_observable(observable_object, 'x509_mapping'))
+            if hasattr(observable_object, 'hashes'):
+                attributes.extend(self._get_attributes_from_observable(observable_object.hashes, 'x509_mapping'))
+        self.handle_import_case(observable, attributes, 'x509')
+
+    ################################################################################
+    ##                         PATTERN PARSING FUNCTIONS.                         ##
+    ################################################################################
+
+    @staticmethod
+    def _fetch_user_account_type_pattern(pattern):
+        for stix_object in pattern:
+            if 'extensions' in stix_object or all(key not in stix_object for key in ('user_id', 'credential', 'type')):
+                return 'user-account', 'user_account_mapping'
+        return 'credential', 'credential_mapping'
+
+    def get_attachment(self, attachment, filename):
+        attribute = {
+            'type': 'attachment',
+            'object_relation': 'attachment',
+            'value': attachment.pop(filename)
+        }
+        data_feature = self._choose_with_priority(attachment, 'file:content_ref.payload_bin', 'artifact:payload_bin')
+        attribute['data'] = attachment.pop(data_feature)
+        return attribute
+
+    def get_attributes_from_pattern(self, pattern, mapping):
+        attributes = []
+        for pattern_part in pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            try:
+                attribute = deepcopy(getattr(stix2misp_mapping, mapping)[pattern_type])
+            except KeyError:
+                print(f'Pattern type not supported at the moment: {pattern_type}', file=sys.stderr)
+                continue
+            attribute['value'] = pattern_value
+            attributes.append(attribute)
+        return attributes
+
+    def get_malware_sample(self, attachment, filename):
+        md5_feature = self._choose_with_priority(attachment, "file:content_ref.hashes.'MD5'", "file:hashes.'MD5'")
+        attribute = {
+            'type': 'malware-sample',
+            'object_relation': 'malware-sample',
+            'value': f'{attachment.pop(filename)}|{attachment.pop(md5_feature)}'
+        }
+        data_feature = self._choose_with_priority(attachment, 'file:content_ref.payload_bin', 'artifact:payload_bin')
+        attribute['data'] = attachment.pop(data_feature)
+        return attribute
+
+    def _handle_file_attachments(self, attachment):
+        attributes = []
+        if any('content_ref' in feature for feature in attachment.keys()):
+            attribute_type = 'attachment'
+            value = attachment['file:name'] if 'file:name' in attachment else 'unknown_filename'
+            if "file:content_ref.hashes.'MD5'" in attachment:
+                attribute_type = 'malware-sample'
+                md5 = attachment.pop("file:content_ref.hashes.'MD5'")
+                value = f'{value}|{md5}'
+            data = self._choose_with_priority(attachment, 'file:content_ref.payload_bin', 'artifact:payload_bin')
+            attribute = {
+                'type': attribute_type,
+                'object_relation': attribute_type,
+                'value': value,
+                'data': attachment.pop(data)
+            }
+            attributes.append(attribute)
+        if 'artifact:payload_bin' in attachment:
+            attribute = {
+                'type': 'attachment',
+                'object_relation': 'attachment',
+                'value': attachment['file:name'],
+                'data': attachment.pop('artifact:payload_bin')
+            }
+            attributes.append(attribute)
+        return attributes
+
+    def parse_as_pattern(self, indicator):
+        attributes = self.get_attributes_from_pattern(indicator.pattern, 'asn_mapping')
+        self.handle_import_case(indicator, attributes, 'asn')
+
+    def parse_domain_ip_port_pattern(self, indicator):
+        attributes = []
+        references = defaultdict(dict)
+        for pattern_part in indicator.pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            if pattern_type not in stix2misp_mapping.domain_ip_mapping:
+                if any(pattern_type.startswith(f'network-traffic:{feature}_ref') for feature in ('src', 'dst')):
+                    feature_type, ref = pattern_type.split(':')[1].split('_')
+                    ref, feature = ref.split('.')
+                    ref = f"{feature_type}_{'0' if ref == 'ref' else ref.strip('ref[]')}"
+                    references[ref].update(self._parse_network_connection_reference(feature_type, feature, pattern_value))
+                continue
+            attribute = deepcopy(stix2misp_mapping.domain_ip_mapping[pattern_type])
+            attribute['value'] = pattern_value
+            attributes.append(attribute)
+        if references:
+            attributes.extend(references.values())
+        object_name = 'ip-port' if 'network-traffic' in indicator.pattern else 'domain-ip'
+        self.handle_import_case(indicator, attributes, object_name)
+
+    def parse_email_message_pattern(self, indicator):
+        attributes = []
+        attachments = defaultdict(dict)
+        for pattern_part in indicator.pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            if pattern_type not in stix2misp_mapping.email_mapping:
+                if pattern_type.startswith('email-message:body_multipart'):
+                    features = pattern_type.split('.')
+                    if len(features) == 3 and features[1] == 'body_raw_ref':
+                        index = features[0].split('[')[1].strip(']') if '[' in features[0] else '0'
+                        key = 'data' if features[2] == 'payload_bin' else 'value'
+                        attachments[index][key] = pattern_value
+                        continue
+                print(f'Pattern type not supported at the moment: {pattern_type}', file=sys.stderr)
+                continue
+            attribute = deepcopy(stix2misp_mapping.email_mapping[pattern_type])
+            attribute['value'] = pattern_value
+            attributes.append(attribute)
+        if attachments:
+            for attachment in attachments.values():
+                attribute = {
+                    'type': 'attachment',
+                    'object_relation': 'screenshot'
+                } if 'data' in attachment else {
+                    'type': 'email-attachment',
+                    'object_relation': 'attachment'
+                }
+                attribute.update(attachment)
+                attributes.append(attribute)
+        self.handle_import_case(indicator, attributes, 'email')
+
+    def parse_file_pattern(self, indicator):
+        attributes = []
+        attachment = {}
+        extensions = defaultdict(lambda: defaultdict(dict))
+        for pattern_part in indicator.pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            if pattern_type in stix2misp_mapping.attachment_types:
+                attachment[pattern_type] = pattern_value.strip("'")
+                continue
+            if pattern_type not in stix2misp_mapping.file_mapping:
+                if 'extensions' in pattern_type:
+                    features = pattern_type.split('.')[1:]
+                    extension_type = features.pop(0).strip("'")
+                    if 'section' in features[0] and features[0] != 'number_of_sections':
+                        index = features[0].split('[')[1].strip(']') if '[' in features[0] else '0'
+                        extensions[extension_type][f'section_{index}'][features[-1].strip("'")] = pattern_value
+                    else:
+                        extensions[extension_type]['.'.join(features)] = pattern_value
+                continue
+            attribute = deepcopy(stix2misp_mapping.file_mapping[pattern_type])
+            attribute['value'] = pattern_value
+            attributes.append(attribute)
+        if any(key.endswith('payload_bin') for key in attachment.keys()):
+            attributes.extend(self._handle_file_attachments(attachment))
+        if attachment:
+            for pattern_type, value in attachment.items():
+                if pattern_type in stix2misp_mapping.file_mapping:
+                    attribute = deepcopy(stix2misp_mapping.file_mapping[pattern_type])
+                    attribute['value'] = value
+                    attributes.append(attribute)
+        if extensions:
+            file_object = self.create_misp_object(indicator, 'file')
+            self.parse_file_extension(file_object, attributes, extensions)
+        else:
+            self.handle_import_case(indicator, attributes, 'file')
+
+    def parse_file_extension(self, file_object, attributes, extensions):
+        for attribute in attributes:
+            file_object.add_attribute(**attribute)
+        if 'windows-pebinary-ext' in extensions:
+            pe_extension = extensions['windows-pebinary-ext']
+            pe_object = MISPObject('pe', misp_objects_path_custom=_misp_objects_path)
+            sections = self._get_sections(pe_extension)
+            self.fill_misp_object_from_dict(pe_object, pe_extension, 'pe_mapping')
+            file_object.add_reference(pe_object.uuid, 'includes')
+            if sections:
+                for section in sections:
+                    section_object = MISPObject('pe-section')
+                    self.fill_misp_object_from_dict(section_object, section, 'pe_section_mapping')
+                    pe_object.add_reference(section_object.uuid, 'includes')
+                    self.misp_event.add_object(**section_object)
+            self.misp_event.add_object(**pe_object)
         self.misp_event.add_object(**file_object)
 
-    def parse_asn_observable(self, objects, uuid, marking=None):
-        attributes = self.attributes_from_asn_observable(objects)
-        self.handle_import_case(attributes, 'asn', marking, uuid)
+    def parse_ip_address_pattern(self, indicator):
+        self.add_attribute_from_indicator(indicator, 'ip-dst')
 
-    def parse_domain_ip_observable(self, objects, uuid, marking=None):
-        attributes = self.attributes_from_domain_ip_observable(objects)
-        self.handle_import_case(attributes, 'domain-ip', marking, uuid)
+    def parse_mac_address_pattern(self, indicator):
+        self.add_attribute_from_indicator(indicator, 'mac-address')
 
-    def parse_domain_ip_port_pattern(self, pattern, marking=None, uuid=None):
-        values = {}
-        for p in pattern:
-            type_, value = p.split('=')
-            values[type_.strip().split(':')[0]] = value.strip().strip('\'')
-        attributes = self.attributes_from_dict(values, domain_ip_mapping, True)
-        self.handle_import_case(attributes, 'domain-ip', marking, uuid)
+    def parse_mutex_pattern(self, indicator):
+        self.add_attribute_from_indicator(indicator, 'mutex')
 
-    def parse_email_observable(self, objects, uuid, marking=None):
-        to_ids = False
-        attributes, message = self.parse_complex_fields_observable_email(objects, to_ids)
-        for m_key, m_value in message.items():
-            if m_key in email_mapping:
-                attributes.append(self.append_email_attribute(m_key, m_value, to_ids))
-        self.handle_import_case(attributes, 'email', marking, uuid)
+    def parse_network_connection_pattern(self, indicator, attributes, references):
+        attributes.extend(self._parse_network_pattern_references(references, 'network_traffic_references_mapping'))
+        self.handle_import_case(indicator, attributes, 'network-connection')
 
-    def parse_email_address_observable(self, objects, uuid, marking=None):
-        mapping = to_attribute_mapping
-        attributes = [{'type': 'email-dst', 'object_relation': 'to', 'to_ids': True, 'value': _object.value} for _object in objects.values()]
-        self.handle_import_case(attributes, 'email', marking, uuid)
+    @staticmethod
+    def _parse_network_pattern_references(references, mapping):
+        attributes = []
+        for feature, reference in references.items():
+            feature = feature.split('_')[0]
+            attribute = {key: value.format(feature) for key, value in getattr(stix2misp_mapping, mapping)[reference['type']].items()}
+            attribute['value'] = reference['value']
+            attributes.append(attribute)
+        return attributes
 
-    def parse_email_address_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.fill_pattern_attributes(pattern_types, pattern_values, email_mapping)
-        self.handle_import_case(attributes, 'email', marking, uuid)
+    def parse_network_socket_pattern(self, indicator, attributes, references, extension):
+        attributes.extend(self._parse_network_pattern_references(references, 'network_traffic_references_mapping'))
+        for key, value in extension.items():
+            if key not in stix2misp_mapping.network_socket_extension_mapping:
+                print(f'Unknown socket extension field in pattern: {key}', file=sys.stderr)
+                continue
+            if key.startswith('is_') and not json.loads(value.lower()):
+                continue
+            attribute = deepcopy(stix2misp_mapping.network_socket_extension_mapping[key])
+            attribute['value'] = key.split('_')[1] if key.startswith('is_') else value
+            attributes.append(attribute)
+        self.handle_import_case(indicator, attributes, 'network-socket')
 
-    def parse_file_observable(self, objects, uuid, marking=None):
-        file, directory = self.split_file_observable(objects)
-        attributes = self.attributes_from_file_observable(file)
-        if directory is not None and directory.path:
-            mapping = file_mapping['path']
-            attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                               'value': directory.path, 'to_ids': False})
-        if hasattr(_object, 'extensions') and 'windows-pebinary-ext' in _object.extensions:
-            self.handle_pe_case(_object.extensions['windows-pebinary-ext'], attributes, uuid)
-        else:
-            self.handle_import_case(attributes, _object._type, marking, uuid)
+    def parse_network_traffic_pattern(self, indicator):
+        attributes = []
+        protocols = []
+        references = defaultdict(dict)
+        extensions = defaultdict(dict)
+        for pattern_part in indicator.pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            if pattern_type in stix2misp_mapping.network_traffic_mapping:
+                attribute = deepcopy(stix2misp_mapping.network_traffic_mapping[pattern_type])
+                attribute['value'] = pattern_value.strip("'")
+                attributes.append(attribute)
+                continue
+            if pattern_type.startswith('network-traffic:protocols['):
+                protocols.append(pattern_value)
+            elif any(pattern_type.startswith(f'network-traffic:{feature}_ref') for feature in ('src', 'dst')):
+                feature_type, ref = pattern_type.split(':')[1].split('_')
+                ref, feature = ref.split('.')
+                ref = f"{feature_type}_{'0' if ref == 'ref' else ref.strip('ref[]')}"
+                references[ref].update({feature: pattern_value})
+            elif pattern_type.startswith('network-traffic:extensions.'):
+                _, extension_type, feature = pattern_type.split('.')
+                extensions[extension_type.strip("'")][feature] = pattern_value
+            else:
+                print(f'Pattern type not supported at the moment: {pattern_type}', file=sys.stderr)
+        if extensions:
+            if 'socket-ext' in extensions:
+                return self.parse_network_socket_pattern(indicator, attributes, references, extensions['socket-ext'])
+            print(f'Unknown network extension(s) in pattern: {", ".join(extensions.keys())}', file=sys.stderr)
+        if protocols and self._required_protocols(protocols):
+            attributes.extend(self.parse_protocols(protocols, 'pattern'))
+            return self.parse_network_connection_pattern(indicator, attributes, references)
+        attributes.extend(self._parse_network_pattern_references(references, 'ip_port_references_mapping'))
+        self.handle_import_case(indicator, attributes, 'ip-port')
 
-    def parse_file_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.attributes_from_file_pattern(pattern_types, pattern_values)
-        if any((attribute['object_relation'] == 'path' for attribute in attributes)):
-            self.object_case_import(attributes, 'file', uuid)
-        else:
-            self.handle_import_case(attributes, 'file', marking, uuid)
+    def parse_process_pattern(self, indicator):
+        attributes = []
+        parent = {}
+        child = defaultdict(set)
+        for pattern_part in indicator.pattern.strip('[]').split(' AND '):
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            if 'parent_' in pattern_type:
+                child[pattern_type.split('.')[-1]].add(pattern_value)
+            elif 'child_' in pattern_type:
+                parent[pattern_type.split('.')[-1]] = pattern_value
+            else:
+                try:
+                    attribute = deepcopy(stix2misp_mapping.process_mapping[pattern_type])
+                except KeyError:
+                    print(f'Pattern type not supported at the moment: {pattern_type}', file=sys.stderr)
+                    continue
+                attribute['value'] = pattern_value
+                attributes.append(attribute)
+        if parent:
+            for key, value in parent.items():
+                if key not in stix2misp_mapping.parent_process_reference_mapping:
+                    print(f'Parent process key from pattern not supported at the moment: {key}', file=sys.stderr)
+                    continue
+                attribute = {'value': value}
+                attribute.update(stix2misp_mapping.parent_process_reference_mapping[key])
+                attributes.append(attribute)
+        if child:
+            for key, values in child.items():
+                if key not in stix2misp_mapping.child_process_reference_mapping:
+                    print(f'Child process key from pattern not supported at the moment: {key}', file=sys.stderr)
+                    continue
+                for value in values:
+                    attribute = {'value': value}
+                    attribute.update(stix2misp_mapping.child_process_reference_mapping[key])
+                    attributes.append(attribute)
+        self.handle_import_case(indicator, attributes, 'process')
 
-    def parse_file_object_observable(self, objects, uuid, marking=None):
-        file, data = self.split_file_observable(objects)
-        attributes = self.attributes_from_file_observable(file, data)
-        if hasattr(file, 'extensions') and 'windows-pebinary-ext' in file.extensions:
-            self.handle_pe_case(file.extensions['windows-pebinary-ext'], attributes, uuid)
-        else:
-            self.handle_import_case(attributes, file._type, marking, uuid)
+    def parse_regkey_pattern(self, indicator):
+        attributes = self.get_attributes_from_pattern(indicator.pattern, 'regkey_mapping')
+        self.handle_import_case(indicator, attributes, 'registry-key')
 
-    def parse_ip_address_observable(self, objects, uuid, marking=None):
-        self.add_attributes_from_observable(objects, 'ip-dst', 'value', uuid, marking)
+    def parse_url_pattern(self, indicator):
+        attributes = self.get_attributes_from_pattern(indicator.pattern, 'url_mapping')
+        self.handle_import_case(indicator, attributes, 'url')
 
-    def parse_ip_address_pattern(self, pattern, marking=None, uuid=None):
-        self.add_attributes_from_pattern('ip-dst', pattern, marking, uuid)
+    def parse_user_account_pattern(self, indicator):
+        attributes = []
+        pattern = indicator.pattern.strip('[]').split(' AND ')
+        object_name, mapping = self._fetch_user_account_type_pattern(pattern)
+        for pattern_part in pattern:
+            pattern_type, pattern_value = self.get_type_and_value_from_pattern(pattern_part)
+            pattern_type = pattern_type.split(':')[1]
+            if pattern_type.startswith('extensions.'):
+                pattern_type = pattern_type.split('.')[-1]
+                if '[' in pattern_type:
+                    pattern_type = pattern_type.split('[')[0]
+                if pattern_type in ('group', 'groups'):
+                    attributes.append({'type': 'text', 'object_relation': 'group', 'value': pattern_value})
+                    continue
+            if pattern_type in getattr(stix2misp_mapping, mapping):
+                attribute = deepcopy(getattr(stix2misp_mapping, mapping)[pattern_type])
+                attribute['value'] = pattern_value
+                attributes.append(attribute)
+        self.handle_import_case(indicator, attributes, object_name)
 
-    def parse_ip_network_traffic_observable(self, objects, uuid, marking=None):
-        attributes, name = self.attributes_from_network_traffic(objects)
-        self.handle_import_case(attributes, name, marking, uuid)
-
-    def parse_ip_port_or_network_socket_observable(self, objects, uuid, marking=None):
-        attributes, name = self.attributes_from_network_traffic(objects)
-        self.handle_import_case(attributes, name, marking, uuid)
-
-    def parse_mac_address_observable(self, objects, uuid, marking=None):
-        self.add_attributes_from_observable(objects, 'mac-address', 'value', uuid, marking)
-
-    def parse_mutex_observable(self, objects, uuid, marking=None):
-        self.add_attributes_from_observable(objects, 'mutex', 'name', uuid, marking)
-
-    def parse_network_socket_observable(self, objects, uuid, marking=None):
-        attributes, name = self.attributes_from_network_traffic(objects)
-        self.handle_import_case(attributes, name, marking, uuid)
-
-    def parse_network_traffic_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.fill_pattern_attributes(pattern_types, pattern_values, network_traffic_mapping)
-        self.handle_import_case(attributes, 'ip-port', marking, uuid)
-
-    def parse_process_observable(self, objects, uuid, marking=None):
-        attributes = self.attributes_from_process_observable(objects)
-        self.handle_import_case(attributes, 'process', marking, uuid)
-
-    def parse_process_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.fill_pattern_attributes(pattern_types, pattern_values, process_mapping)
-        self.handle_import_case(attributes, 'process', marking, uuid)
-
-    def parse_regkey_observable(self, objects, uuid, marking=None):
-        _object = objects['0']
-        attributes = self.attributes_from_regkey_observable(_object)
-        self.handle_import_case(attributes, 'registry-key', marking, uuid)
-
-    def parse_regkey_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, pattern_values = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.fill_pattern_attributes(pattern_types, pattern_values, regkey_mapping)
-        self.handle_import_case(attributes, 'registry-key', marking, uuid)
-
-    def parse_url_observable(self, objects, uuid, marking=None):
-        self.add_attributes_from_observable(objects, 'url', 'value', uuid, marking)
-
-    def parse_url_pattern(self, pattern, marking=None, uuid=None):
-        self.add_attributes_from_pattern('url', pattern, marking, uuid)
-
-    def parse_url_object_observable(self, objects, uuid, marking=None):
-        attributes = self.attributes_from_url_observable(objects)
-        self.handle_import_case(attributes, 'url', marking, uuid)
-
-    def parse_user_account_observable(self, observable, uuid, marking=None):
-        attributes = self.attributes_from_user_account_observable(observable)
-        name = self.__define_user_account_name(attributes)
-        self.handle_import_case(attributes, name, marking, uuid)
-
-    def parse_user_account_pattern(self, pattern, marking, uuid):
-        attributes = self.attributes_from_user_account_pattern(pattern)
-        name = self.__define_user_account_name(attributes)
-        self.handle_import_case(attributes, name, marking, uuid)
-
-    def parse_x509_observable(self, objects, uuid, marking=None):
-        attributes = self.attributes_from_x509_observable(objects)
-        self.handle_import_case(attributes, 'x509', marking, uuid)
-
-    def parse_x509_pattern(self, pattern, marking=None, uuid=None):
-        pattern_types, attribute_types = self.get_types_and_values_from_pattern(pattern)
-        attributes = self.fill_pattern_attributes(pattern_types, attribute_types, x509_mapping)
-        self.handle_import_case(attributes, 'x509', marking, uuid)
+    def parse_x509_pattern(self, indicator):
+        attributes = self.get_attributes_from_pattern(indicator.pattern, 'x509_mapping')
+        self.handle_import_case(indicator, attributes, 'x509')
 
     ################################################################################
     ##                             UTILITY FUNCTIONS.                             ##
     ################################################################################
 
-    def add_stix2_pattern_object(self, pattern, uuid):
-        attribute = {'type': 'stix2-pattern', 'object_relation': 'stix2-pattern', 'value': pattern}
-        misp_object = {'name': 'stix2-pattern', 'meta-category': 'stix2-pattern',
-                       'Attribute': [self.version_attribute, attribute],
-                       'uuid': uuid}
-        self.misp_event.add_object(**misp_object)
+    def add_attribute_from_indicator(self, indicator, attribute_type):
+        patterns = indicator.pattern.strip('[]').split(' AND ')
+        if len(patterns) == 1:
+            _, value = self.get_type_and_value_from_pattern(patterns[0])
+            attribute = MISPAttribute()
+            attribute.from_dict(**{
+                'uuid': indicator.id.split('--')[1],
+                'type': attribute_type,
+                'value': value,
+                'to_ids': True
+            })
+            attribute.update(self.parse_timeline(indicator))
+            self.misp_event.add_attribute(**attribute)
+        else:
+            tmp_attribute = self.parse_timeline(indicator)
+            for pattern in patterns:
+                _, value = self.get_type_and_value_from_pattern(pattern)
+                attribute = MISPAttribute()
+                attribute.from_dict(**{
+                    'type': attribute_type,
+                    'value': value,
+                    'to_ids': True
+                })
+                attribute.update(tmp_attribute)
+                self.misp_event.add_attribute(**attribute)
 
-    @staticmethod
-    def create_misp_object(attributes, name, uuid=None):
-        misp_object = MISPObject(name, misp_objects_path_custom=_MISP_objects_path)
-        if uuid is not None:
-            misp_object.uuid = uuid
-        for attribute in attributes:
-            misp_object.add_attribute(**attribute)
+    def add_attribute_from_observable(self, observable, attribute_type, feature):
+        attribute = MISPAttribute()
+        attribute.from_dict(**{
+            'uuid': observable.id.split('--')[1],
+            'type': attribute_type,
+            'value': getattr(observable.objects['0'], feature),
+            'to_ids': False
+        })
+        attribute.update(self.parse_timeline(observable))
+        self.misp_event.add_attribute(**attribute)
+
+    def add_attributes_from_observable(self, observable_objects, attribute_type, feature):
+        attribute = {'type': attribute_type, 'to_ids': False}
+        for observable_object in observable_objects.values():
+            tmp_attribute = {'value': getattr(observable_object, feature)}
+            tmp_attribute.update(attribute)
+            self.misp_event.add_attribute(**tmp_attribute)
+
+    def create_misp_object(self, stix_object, name=None):
+        misp_object = MISPObject(name if name is not None else stix_object.type,
+                                 misp_objects_path_custom=_misp_objects_path)
+        misp_object.uuid = stix_object.id.split('--')[1]
+        misp_object.update(self.parse_timeline(stix_object))
         return misp_object
 
     @staticmethod
-    def __define_user_account_name(attributes):
-        if len(attributes) == 2:
-            relations = (attribute['type'] for attribute in attributes)
-            if 'user_id' in relations and 'credential' in relations:
-                return 'credential'
-        return 'user-account'
+    def _get_sections(pe_extension):
+        sections = [feature for feature in pe_extension.keys() if feature.startswith('section_')]
+        return (pe_extension.pop(feature) for feature in sections)
 
     @staticmethod
-    def fill_observable_attributes(stix_object, object_mapping):
-        attributes = []
-        for o_key, o_value in stix_object.items():
-            if o_key in object_mapping:
-                mapping = object_mapping[o_key]
-                attributes.append({'type': mapping.get('type'), 'object_relation': mapping.get('relation'),
-                                   'value': o_value, 'to_ids': False})
-        return attributes
+    def get_type_and_value_from_pattern(pattern):
+        try:
+            pattern_type, pattern_value = pattern.split(' = \'')
+        except ValueError:
+            pattern_type, pattern_value = pattern.split('=')
+        return pattern_type, pattern_value.strip("'")
 
-    def fill_object_attributes_observable(self, misp_object, mapping_dict, stix_object):
-        for stix_type, value in stix_object.items():
-            if stix_type in mapping_dict:
-                mapping = mapping_dict[stix_type]
-                misp_object.add_attribute(**{'type': mapping['type'], 'object_relation': mapping['relation'],
-                                             'value': value, 'to_ids': False})
-
-    @staticmethod
-    def fill_pattern_attributes(types, values, object_mapping):
-        attributes = []
-        for type_, value in zip(types, values):
-            try:
-                mapping = object_mapping[type_]
-                attributes.append({'type': mapping['type'], 'object_relation': mapping['relation'],
-                                   'to_ids': True, 'value': value.replace('\\\\', '\\')})
-            except KeyError:
-                continue
-        return attributes
-
-    @staticmethod
-    def get_types_and_values_from_pattern(pattern):
-        types = []
-        values = []
-        for p in pattern:
-            try:
-                type_, value = p.split('=')
-            except ValueError:
-                try:
-                    type_, value = p.split(' = ')
-                except ValueError:
-                    type_, value = p.split(' = \'')
-            types.append(type_.strip())
-            values.append(value.strip().strip('\''))
-        return types, values
-
-    def handle_import_case(self, attributes, name, marking, uuid):
+    def handle_import_case(self, stix_object, attributes, name):
         if len(attributes) == 1:
-            attribute = attributes[0]
-            attribute = {field: attribute[field] for field in self.single_attribute_fields if attribute.get(field)}
-            attribute['uuid'] = uuid
-            if marking:
-                attribute = self.add_tag_in_attribute(attribute, marking)
-            self.handle_single_attribute(attribute, uuid=uuid)
+            attribute = {field: attributes[0][field] for field in self._single_attribute_fields if attributes[0].get(field)}
+            attribute['uuid'] = stix_object.id.split('--')[1]
+            attribute.update(self.parse_timeline(stix_object))
+            self.misp_event.add_attribute(**attribute)
         else:
-            self.object_case_import(attributes, name, uuid)
-
-    def object_case_import(self, attributes, name, uuid):
-        misp_object = self.create_misp_object(attributes, name, uuid)
-        if uuid is not None and uuid in self.relationship:
-            self.handle_object_relationship(misp_object, uuid)
-        self.misp_event.add_object(**misp_object)
-
+            misp_object = self.create_misp_object(stix_object, name)
+            for attribute in attributes:
+                misp_object.add_attribute(**attribute)
+            self.misp_event.add_object(**misp_object)
 
     @staticmethod
-    def parse_marking(marking):
-        marking_type = marking.definition_type
-        if marking_type == 'tlp':
-            return "{}:{}".format(marking_type, getattr(marking.definition, marking_type))
+    def _parse_observable_types(observable_objects):
+        types = {observable_object._type for observable_object in observable_objects.values()}
+        return tuple(sorted(types))
+
+    @staticmethod
+    def _parse_pattern_types(pattern):
+        types = {part.split('=')[0].split(':')[0] for part in pattern}
+        return tuple(sorted(types))
+
+    @staticmethod
+    def _required_protocols(protocols):
+        protocols = tuple(protocol.upper() for protocol in protocols)
+        if any(protocol not in ('TCP', 'IP') for protocol in protocols):
+            return True
+        return False
 
 
 def from_misp(stix_objects):
-    for o in stix_objects:
-        if o._type == "report" and 'misp:tool="misp2stix2"' in o.get('labels'):
+    for stix_object in stix_objects:
+        if stix_object['type'] == "report" and 'misp:tool="misp2stix2"' in stix_object.get('labels', []):
             return True
     return False
 
+
 def main(args):
-    filename = os.path.join(os.path.dirname(args[0]), args[1])
+    filename = Path(os.path.dirname(args[0]), args[1])
     with open(filename, 'rt', encoding='utf-8') as f:
         event = stix2.parse(f.read(), allow_custom=True, interoperability=True)
-    stix_version = 'STIX {}'.format(event.get('spec_version'))
     stix_parser = StixFromMISPParser() if from_misp(event.objects) else ExternalStixParser()
-    stix_parser.load_data(filename, stix_version, event, args[2:])
-    stix_parser.handler()
-    stix_parser.saveFile()
+    stix_parser.handler(event, filename, args[2:])
+    stix_parser.save_file()
     print(1)
 
-if __name__ == "__main__":
+
+if __name__ == '__main__':
     main(sys.argv)
